### 2019-01-15 #1175 @whedon
[PRE REVIEW]: QMix: A Python package for simulating the quasiparticle tunneling currents in SIS junctions
**Submitting author:** @garrettj403 (<a href="http://orcid.org/0000-0002-5509-5006">John Garrett</a>)
**Repository:** <a href="https://github.com/garrettj403/QMix/" target ="_blank">https://github.com/garrettj403/QMix/</a>
**Version:** v1.0.0
