### 2019-01-16 e48d15b
Expand contribution instructions [ci skip]
