+++ CONTRIBUTING.md
@@ -1,18 +1,28 @@
-Contributing to QMix
-====================
+How to contribute to QMix 
+=========================
 
-Please feel free to contribute to QMix! 
+We welcome contributions to QMix via pull requests:
+1. fork the QMix repository to your own GitHub profile, 
+2. make any changes that you like, and then 
+3. create a pull request to merge your changes back into the main QMix repository. 
 
-To contribute, fork the QMix repository to your own profile, make any changes that you like, and then create a pull request to merge your changes back into the main QMix repository. I'll do my best to review your changes in a timely manner.
+I'll do my best to review your changes in a timely manner.
+
+**Note:** For information on how to download and install QMix, please see the [QMix website](https://garrettj403.github.io/QMix/setup.html).
 
 Bugs 
 ----
 
-If you find any bugs or if you would like a specific feature added, please create a new issue through the [GitHub Issue tracker](https://github.com/garrettj403/QMix/issues).
+If you find any bugs, please either create a new issue through the [GitHub Issue tracker](https://github.com/garrettj403/QMix/issues) with the ``bug`` label, or fix the issue yourself via a pull request (thank you in advance!)
+
+New Features
+------------
+
+If you would like a new feature added, please create a new issue through the [GitHub Issue tracker](https://github.com/garrettj403/QMix/issues) with the ``enhancement`` label. I can likely advise how to implement the new feature, but there is no guarantee that the feature will be added.
 
 Problems Running or Installing QMix
 -----------------------------------
 
 If you have any problems getting QMix to work, first read through the [installation instructions](https://garrettj403.github.io/QMix/setup.html#installing-qmix-using-pip) online and then make sure you are using the [virtual environment](https://garrettj403.github.io/QMix/setup.html#configuring-python) (``QMix/environment.yml``) that I included with QMix.
 
-If you are still having problems, please email me at garrettj403@gmail.com. I can also help you setup your simulations.
+If you are still having problems, please either [create a new issue](https://github.com/garrettj403/QMix/issues) or email me at garrettj403@gmail.com. I can also help you setup your simulations.
