+++ _sources/setup.rst.txt
@@ -10,8 +10,8 @@ The easiest way to install QMix is using ``pip``:
 
    pip install QMix
 
-Installing QMix from GitHub
----------------------------
+Downloading QMix from GitHub
+----------------------------
 
 If you want the latest version, you can download QMix from `GitHub`_. In order to clone the repository, open a terminal window, navigate to the directory in which you would like to put QMix, and then run:
 
@@ -32,7 +32,7 @@ where ``<path-to-qmix>`` is the path to the ``QMix/`` directory on your computer
 Configuring Python
 ------------------
 
-QMix is intended for Python version 3.7. If you don't want to change your Python installation or in case you are missing any packages, I have included a virtual environment file that contains all of the neccesary packages. If you are using the Anaconda package manager, you can run:
+QMix is written using Python version 3.7. If you don't want to change your Python installation or in case you are missing any packages, I have included a virtual environment file that contains all of the neccesary packages. If you are using the Anaconda package manager, you can run:
 
 .. code-block:: bash
 
