### 2019-01-16 a0e9090
Add Python version 3.5 and 3.6 to Travis tests
