+++ .travis.yml
@@ -1,7 +1,16 @@
+# Parameters for Travis CI tests
+# These are automated tests that run after each commit
+# see: https://docs.travis-ci.com/user/languages/python/
 language: python
 python:
-  - "3.7-dev"
+  - "3.5"
+  - "3.5-dev"  # 3.5 development branch
+  - "3.6"
+  - "3.6-dev"  # 3.6 development branch
+  - "3.7-dev"  # 3.7 development branch
+# command to install dependencies
 install:
   - pip install -r requirements.txt
   - pip install .
+# command to run tests
 script: pytest
