### 2019-01-27 #1175 @labarba
This submission is pretty far outside of my domains of even fuzzy understanding, but I'm going to try some gut-feeling pings to some registered JOSS reviewers. Maybe signal processing?
—Could this be something you can review for JOSS, @josephhardinee?
