### 2019-01-29 #1175 @josephhardinee
Sure. It is pretty far outside of my area of expertise, but
if @FaustinCarter can handle domain specific I can help look at some of the
more general portions.

On Mon, Jan 28, 2019 at 3:09 PM Lorena A. Barba <notifications@github.com>
wrote:

> Just pinged @josephhardinee <https://github.com/josephhardinee> by email
> to see if he can be our second reviewer. Stay tuned.
>
