### 2019-01-30 #1175 @garrettj403
Thank you @josephhardinee and @FaustinCarter for agreeing to review QMix!

@labarba, I've sent a few emails around to try to find a third reviewer. I'll let you know as soon as I hear back.
