### 2019-02-04 #1175 @labarba
👋  @PaulKGrimes — We hear you are willing to provide a review for this submission to JOSS? (Please confirm.) We have two other reviewers, but since the subject of the paper is far out of my areas of expertise, I would welcome another pair of eyes on this.
