### 2019-02-04 #1175 @PaulKGrimes
Hi Lorena,

Yes, I'm absolutely willing to review this submission.  I've been working
on SIS mixer simulation on and off in Python and C++ for the past 16 years,
so hopefully I can provide useful feedback.

Paul Grimes
--
Submillimeter Array/Receiver Laboratory
Center for Astrophysics | Harvard & Smithsonian
+1 (617) 496-7640

On Mon, Feb 4, 2019 at 2:12 PM Lorena A. Barba <notifications@github.com>
wrote:

> 👋 @PaulKGrimes <https://github.com/PaulKGrimes> — We hear you are
> willing to provide a review for this submission to JOSS? (Please confirm.)
> We have two other reviewers, but since the subject of the paper is far out
> of my areas of expertise, I would welcome another pair of eyes on this.
>
