### 2019-02-05 #1231 @FaustinCarter
I've had a go at installing it and had some issues. I'll drop them here now so they can be addressed in parallel with other things.

Attempts at installation:

1. I can't use the included environment.yml file on Windows. I'm not sure if this is Windows-specific, but the yml file specifies the exact hash to use for each package version and those don't match the ones I see when I do `conda search foo`. Recommendation is to just specify version numbers and let conda sort out which system-specific repo it should pull from.
2. Straight-up pip install in fresh conda environment (with just python 3.7) failed also because of error trying to read in the long description in setup.py from README.md, which has some non-readable characters (i.e. images?). This suggests the code on PyPI is not the same as the latest code in master on github. However, they both have the same version number, 1.0.0. Recommend bumping the version to 1.0.1 and reuploading to PyPI (and also creating a new release in github to trigger a Zenodo update).
3. Trying pip install -e from source... Success! So the current code installs fine after a download and local build.

I also ran the tests, which all passed on the first go-around. You might mention in the documentation for the tests that pytest is a dependency that needs to be separately installed in order to use.

Another issue that I see right away is that the documentation doesn't have an API section that contains all the class and function docs. I scanned through the major class docstrings and they appear to be generally in good shape, so I think you can get this most of the way there with Sphinx's autodoc feature.

OK, that's all for now. More as I work through the code base. Once the installation issues are fixed I'll tick that box above in the review.
