### 2019-02-05 64817f8
Change virtual environment to only include necessary pacakges, don't include versions or build codes [ci skip]
