+++ environment.yml
@@ -1,124 +1,0 @@
-name: qmix
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - nbsphinx=0.3.5=py_0
-  - alabaster=0.7.12=py37_0
-  - appnope=0.1.0=py37_0
-  - asn1crypto=0.24.0=py37_0
-  - babel=2.6.0=py37_0
-  - backcall=0.1.0=py37_0
-  - blas=1.0=mkl
-  - bleach=3.0.2=py37_0
-  - ca-certificates=2018.03.07=0
-  - certifi=2018.10.15=py37_0
-  - cffi=1.11.5=py37h6174b99_1
-  - chardet=3.0.4=py37_1
-  - conda=4.5.11=py37_0
-  - conda-env=2.6.0=1
-  - cryptography=2.3.1=py37hdbc3d79_0
-  - cycler=0.10.0=py37_0
-  - dbus=1.13.2=h760590f_1
-  - decorator=4.3.0=py37_0
-  - docutils=0.14=py37_0
-  - entrypoints=0.2.3=py37_2
-  - expat=2.2.6=h0a44026_0
-  - freetype=2.9.1=hb4e5f40_0
-  - gettext=0.19.8.1=h15daf44_3
-  - glib=2.56.2=hd9629dc_0
-  - icu=58.2=h4b95b61_1
-  - idna=2.7=py37_0
-  - imagesize=1.1.0=py37_0
-  - intel-openmp=2019.0=118
-  - ipykernel=5.0.0=py37h39e3cac_0
-  - ipython=7.0.1=py37h39e3cac_0
-  - ipython_genutils=0.2.0=py37_0
-  - ipywidgets=7.4.2=py37_0
-  - jedi=0.13.1=py37_0
-  - jinja2=2.10=py37_0
-  - jpeg=9b=he5867d9_2
-  - jsonschema=2.6.0=py37_0
-  - jupyter=1.0.0=py37_7
-  - jupyter_client=5.2.3=py37_0
-  - jupyter_console=6.0.0=py37_0
-  - jupyter_core=4.4.0=py37_0
-  - kiwisolver=1.0.1=py37h0a44026_0
-  - libcxx=4.0.1=h579ed51_0
-  - libcxxabi=4.0.1=hebd6815_0
-  - libedit=3.1.20170329=hb402a30_2
-  - libffi=3.2.1=h475c297_4
-  - libgfortran=3.0.1=h93005f0_2
-  - libiconv=1.15=hdd342a3_7
-  - libpng=1.6.34=he12f830_0
-  - libsodium=1.0.16=h3efe00b_0
-  - llvmlite=0.25.0=py37h8c7ce04_0
-  - markupsafe=1.0=py37h1de35cc_1
-  - matplotlib=3.0.0=py37h54f8f79_0
-  - mistune=0.8.4=py37h1de35cc_0
-  - mkl=2019.1=144
-  - mkl_fft=1.0.6=py37h27c97d8_0
-  - mkl_random=1.0.2=py37h27c97d8_0
-  - nbconvert=5.3.1=py37_0
-  - nbformat=4.4.0=py37_0
-  - ncurses=6.1=h0a44026_0
-  - notebook=5.7.0=py37_0
-  - numba=0.40.0=py37h6440ff4_0
-  - numpy=1.15.4=py37hacdab7b_0
-  - numpy-base=1.15.4=py37h6575580_0
-  - openssl=1.0.2p=h1de35cc_0
-  - packaging=18.0=py37_0
-  - pandoc=2.2.3.2=0
-  - pandocfilters=1.4.2=py37_1
-  - parso=0.3.1=py37_0
[…]
