### 2019-02-05 c56e09b
Remove package versions from requirements.txt
