### 2019-02-05 b434fe1
Fix description and update version number to v1.0.1.
