### 2019-02-05 b1a4125
Set minimum Python version to 3.5 in virtual environment
