+++ environment-qmix.yaml
@@ -3,7 +3,7 @@ channels:
   - defaults
   - conda-forge
 dependencies:
-  - python=3.7.*
+  - python>=3.5
   # QMix dependencies (same as requirements.txt)
   - matplotlib
   - numba
