### 2019-02-05 eb263d3
Remove non-ASCII characters from README file
