### 2019-02-05 #1231 @garrettj403
Hi @FaustinCarter, thank you for your feedback.

As Paul mentioned, I have updated the environment file so that it should work on Windows now. I was previously using `conda env export > environment.yml` to generate the environment.yml file, but this doesn't work across different platforms. I have rewritten the environment to only include the necessary packages without specifying versions or build codes.

I've also fixed the issue with the `setup.py` file. I was previously reading in `README.md` in order to generate the description for `setup.py`, but `README.md` contained some non-ASCII characters. I have now rewritten the description so that this doesn't happen anymore.
