+++ setup.py
@@ -8,7 +8,7 @@ setup(
     description = ("Simulate SIS mixer operation"),
     license = "GPL v3",
     keywords = "SIS mixers, radio astronomy, superconducting detectors, terahertz instrumentation, Python",
-    url = "https://github.com/garrettj403/QMix/archive/v1.0.0.tar.gz",
+    url = "https://garrettj403.github.io/QMix/",
     packages=find_packages(),
     install_requires=[
         'matplotlib',
