### 2019-02-08 9a5f377
Update docstrings for Sphinx documentation [ci skip]
