+++ qmix/mathfn/ivcurve_models.py
@@ -1,4 +1,4 @@
-"""Models to generate DC I-V curves.
+""" This sub-module contains DC I-V curve models.
 
 These models range from simple (e.g., 'perfect' and 'polynomial') to 
 complex (e.g., 'exponential' and 'expanded').
@@ -11,13 +11,15 @@ import numpy as np
 # I-V curve models -----------------------------------------------------------
 
 def perfect(voltage):
-    """Perfect I-V curve.
+    """Generate a perfect I-V curve.
 
-    This is the ideal I-V curve. The current is equal to 0 when |Vb| < 1, and 
-    equal to 0.5 when |Vb| = 1 (where Vb is the bias voltage). Otherwise, the 
-    normalized current is equal to the normalized voltage (normalized to Igap 
-    and Vgap, respectively, where Igap=Vgap/Rn is the gap current, Vgap is the
-    gap voltage, and Rn is the normal resistance).
+    This is the ideal I-V curve. The current is equal to 0 when the normalized 
+    bias voltage is less than 1, equal to 0.5 when the normalized bias voltage
+    is equal to 1, and equal to the normalized bias voltage otherwise.
+    
+    The current and voltage are normalized to Igap and Vgap, respectively, 
+    where Igap=Vgap/Rn is the gap current, Vgap is the gap voltage, and Rn is 
+    the normal resistance.
 
     Args:
         voltage (ndarray): normalized bias voltage
@@ -41,10 +43,15 @@ def perfect(voltage):
 
 
 def perfect_kk(voltage, max_kk=100.):
-    """Kramers-Kronig transform of the perfect I-V curve.
+    """Generate Kramers-Kronig transform of the perfect I-V curve.
+
+    This is an analytic solution.
 
     Args:
         voltage (ndarray): normalized bias voltage
+
+    Keyword Args:
+        max_kk (float): if output is NaN, use this max value instead
 
     Returns:
         ndarray: kk-transform of perfect i-v curve
@@ -76,9 +83,9 @@ def perfect_kk(voltage, max_kk=100.):
 
 
 def polynomial(voltage, order=50):
-    """Polynomial I-V curve model.
+    """Generate the polynomial I-V curve model.
 
-    From Kennedy's thesis (1999).
+    From Kennedy (1999) [see full references in online docs].
 
     Args:
         voltage (ndarray): normalized bias voltage
@@ -95,14 +102,10 @@ def polynomial(voltage, order=50):
 
 
 def exponential(voltage, vgap=2.8e-3, rn=14., rsg=300., agap=4e4, model='fixed'):
-    """The exponential I-V curve model that is used in Chalmers papers.
+    """The exponential I-V curve model that is used in some papers from 
+    Chalmers.
 
-    Reference: 
-        H. Rashid, S. Krause, D. Meledin, V. Desmaris, A. Pavolotsky, and 
-        V. Belitsky, "Frequency Multiplier Based on Distributed 
-        Superconducting Tunnel Junctions: Theory, Design, and 
-        Characterization," IEEE Trans. Terahertz Sci. Technol., pp. 1-13, 
-        2016.
+    From Rashid et al. (2016) [see full references in online docs].
 
     Note:
         - The equation from this paper will result in an I-V curve that has a
@@ -110,12 +113,14 @@ def exponential(voltage, vgap=2.8e-3, rn=14., rsg=300., agap=4e4, model='fixed')
         - The normal resistance will also be slightly lower than it is
           supposed to be.
         - I fixed this model. This model can be selected by setting 
-          model='fixed'. The original model can be selected by 
-          setting model='original'.
+          ``model='fixed'``. The original model can be selected by 
[…]
