+++ qmix/mathfn/kktrans.py
@@ -1,24 +1,26 @@
-"""Generate Kramers-Kronig transform.
+""" This sub-module contains functions to generate the Kramers-Kronig 
+transform of DC I-V data.
+
+Used to find the real component of the response function from the DC I-V 
+curve.
 
 """
 
 import numpy as np
 from scipy.signal import hilbert
 
-# Functions -------------------------------------------------------------------
 
 def kk_trans(v, i, n=50):
-    """Kramers-Kronig transform.
+    """Calculate the Kramers-Kronig transform from DC I-V data.
 
-    Used to find the real component of the response function from the dc I-V
-    curve.
+    Note: 
 
-    Note: v spacing must be constant!
+        Voltage spacing must be constant!
 
     Args:
-        v (ndarray): normalized voltage (dc i-v curve)
-        i (ndarray): normalized current (dc i-v curve)
-        n (int): padding for hilbert transform
+        v (ndarray): normalized voltage (DC I-V curve)
+        i (ndarray): normalized current (DC I-V curve)
+        n (int): padding for Hilbert transform
 
     Returns:
         ndarray: kk transform
@@ -27,7 +29,7 @@ def kk_trans(v, i, n=50):
 
     npts = v.shape[0]
 
-    # Ensure v has even spacing
+    # Ensure v has (roughly) even spacing
     assert np.abs((v[1] - v[0]) - (v[1:] - v[:-1])).max() < 1e-5
 
     # Subtract v to make kk defined at v=infinity
@@ -38,19 +40,23 @@ def kk_trans(v, i, n=50):
 
 
 def kk_trans_trapz(v, i):
-    """Kramers-Kronig transform using simple trapezoidal summation.
+    """Calculate the Kramers-Kronig transform using a simple trapezoidal 
+    summation.
 
-    Used to find the real component of the response function from the dc I-V
-    curve.
+    This function isn't really used anymore, but it is nice to use it to 
+    compare against qmix.mathfn.kktrans.kk_trans.
 
-    Note: v spacing must be constant!
+    Note: 
 
-    This function is (much!) slower than the hilbert transform version. It also
-    has problems with how it is integrated around the singularity.
+        Voltage spacing must be constant!
+
+    This function is (much!) slower than the Hilbert transform version (i.e., 
+    qmix.mathfn.kktrans.kk_trans). It also has problems with how it is 
+    calculated around the singularity.
 
     Args:
-        v (ndarray): dc i-v curve, normalized voltage
-        i (ndarray): dc i-v curve, normalized current
+        v (ndarray): normalized voltage (DC I-V curve)
+        i (ndarray): normalized current (DC I-V curve)
 
     Returns:
         ndarray: kk transform
@@ -60,6 +66,7 @@ def kk_trans_trapz(v, i):
     # Ensure v has even spacing
     assert ((v[1:] - v[:-1]) - (v[1] - v[0])).max() < 1e-5
 
+    # very crude integration
     ikk = []
     for a in range(np.alen(v)):
         v_prime, i_prime = np.delete(v, a), np.delete(i, a)
