+++ qmix/exp/__init__.py
@@ -1,3 +1,6 @@
+""" This sub-package contains functions for importing and analyzing 
+experimental data. """
+
 from qmix.exp.exp_data import (RawData, RawData0,
                                plot_overall_results, plot_if_spectrum,
                                check_iv_if_matching, generate_filename_from_f)
