+++ qmix/mathfn/__init__.py
@@ -1,1 +1,8 @@
+""" This sub-package contains mathematical functions. 
+
+This includes numerical differentiation, I-V curve models, Kramers-Kronig
+transforms and various filters.
+
+"""
+
 from qmix.mathfn.misc import slope, slope_span_n

[…]
