### 2019-02-08 9501063
Add function to automatically generate documentation for QMix [ci skip]
