+++ _sources/important_modules.rst.txt
@@ -1,37 +1,0 @@
-Key Modules
-===========
-
-Embedding Circuit -- :mod:`qmix.circuit`
------------------------------------------------------
-
-.. currentmodule:: qmix.circuit
-
-.. automodule:: qmix.circuit
-   :members:
-
-Response Function -- :mod:`qmix.respfn`
-----------------------------------------------
-
-.. currentmodule:: qmix.respfn
-
-.. automodule:: qmix.respfn
-   :members:
-   :inherited-members:
-
-Harmonic Balance -- :mod:`qmix.harmonic_balance`
----------------------------------------------------------
-
-.. currentmodule:: qmix.harmonic_balance
-
-.. automodule:: qmix.harmonic_balance
-   :members:
-
-QP Tunnelling  Current -- :mod:`qmix.qtcurrent`
---------------------------------------------------
-
-.. currentmodule:: qmix.qtcurrent
-
-.. automodule:: qmix.qtcurrent
-   :members:
-
-
