+++ _sources/references.rst.txt
@@ -1,0 +1,22 @@
+References
+==========
+
+.. note:: For more information on QMix, see Garrett (2018).
+
+.. note:: For more information on multi-tone spectral domain analysis (MTSDA), see Kittara (2002); Kittara, Winthington & Yassin (2007); or Withington, Kittara & Yassin (2003).
+
+Garrett, J. D. (2018). `A 230 GHz Focal Plane Array Using a Wide IF Bandwidth SIS Receiver <https://ora.ox.ac.uk/objects/uuid:d47fbf3b-1cf3-4e58-be97-767b9893066e>`_ (DPhil). University of Oxford.
+
+Kennedy, P. R. (1999) *Superconducting Devices for Millimeter Wavelength Astronomy* (DPhil). University of Cambridge.
+
+Kittara, P. (2002). *The Development of a 700 GHz SIS Mixer with Nb Finline Devices: Nonlinear Mixer Theory, Design Techniques and Experimental Investigation* (DPhil). University of Cambridge.
+
+Kittara, P., Withington, S., & Yassin, G. (2007). `Theoretical and numerical analysis of very high harmonic superconducting tunnel junction mixers. <https://aip.scitation.org/doi/10.1063/1.2424407>`_ Journal of Applied Physics, 101(2), 024508.
+
+Rashid, H., Krause, S., Meledin, D., Desmaris, V., Pavolotsky, A., and Belitsky, V. (2016). *Frequency Multiplier Based on Distributed Superconducting Tunnel Junctions: Theory, Design, and Characterization.* IEEE Trans. Terahertz Sci. Technol., 1-13.
+
+Withington, S., Kittara, P., & Yassin, G. (2003). `Multitone quantum simulations of saturating tunnel junction mixers. <http://aip.scitation.org/doi/10.1063/1.1576515>`_ Journal of Applied Physics, 93(12), 9812–9822.
+
+Withington 1995
+
+Skalare 1989
