+++ _sources/multi-tone.ipynb.txt
@@ -4,7 +4,7 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "# Simulating an SIS mixer\n",
+    "# Example #2: Simulating an SIS mixer\n",
     "\n",
     "- This is a multi-tone simultion (i.e., heterodyne mixing)\n",
     "    - The input consists of 3 signals\n",
@@ -184,12 +184,14 @@
     },
     {
      "data": {
-      "image/png": "iVBORw0KGgoAAAANSUhEUgAAAZYAAAEPCAYAAABhkeIdAAAABHNCSVQICAgIfAhkiAAAAAlwSFlzAAALEgAACxIB0t1+/AAAADl0RVh0U29mdHdhcmUAbWF0cGxvdGxpYiB2ZXJzaW9uIDIuMi4yLCBodHRwOi8vbWF0cGxvdGxpYi5vcmcvhp/UCwAAIABJREFUeJzs3Xd4VFX6wPHvS0ihBpAAoSiCgKAUAVEEkeZSRMoi9oItIqKyYgFZFV2xr7rqovJzFXVtWCkGkBYRy9J778RQTAKBkJ45vz/OJCZhQibJ1OT9PM88M3Pnzr1vbpJ559xz7nnFGINSSinlKVX8HYBSSqmKRROLUkopj9LEopRSyqM0sSillPIoTSxKKaU8ShOLUkopjwqoxCIizURkqYhsFZHNIvKgi3VERN4QkV0iskFEOvsjVqWUUq5V9XcAReQAE4wxa0SkFrBaRBYaY7YUWGcQ0Mp5uwR423mvlFIqAARUi8UYc8gYs8b5+CSwFWhSZLVhwEfG+g2oIyLRPg5VKaVUMQIqsRQkIs2Bi4D/FXmpCXCwwPN4Tk8+Siml/CTQToUBICI1ga+B8caYE0VfdvGW0+alEZEYIAYgIiKiy9lnn+3xOD3N4XBQpUrA5vp8GqfnBEOMoHF6WrDEuWPHjkRjTFSp32iMCagbEAosAB4q5vV3gRsKPN8ORJ9pm61btzbBYOnSpf4OwS0ap+cEQ4zGaJyeFixxAqtMGT7HAypliogA/wG2GmNeLWa12cCtztFhlwIpxphDPgtSKaXUGQXaqbAewC3ARhFZ51z2OHA2gDHmHSAWGAzsAtKA2/0Qp1JKqWIEVGIxxizHdR9KwXUMcJ9vIlKB4vjx4xw6VLqGaWRkJFu3bvVSRJ4RDDFC6eOMiIigadOmhIaGejEqFagCKrEoVZzExESaN29OtWrV3H7PyZMnqVWrlhejKr9giBFKF6cxhqSkJOLj4zn33HO9HJkKRAHVx6JUcbKzs4mIiPB3GMoNIsJZZ51FRkaGv0NRfqKJRQUNO7ZDBQP9XQU/U47qwppYlFJKFbJ//346duxY5vdrH4tSxVi5ciXz5s0jJyeHZ555xt/hKOV1OTk5vPzyy0ydOlVbLEp50oABA0hJSWHWrFk88cQTREVFnTYi7ZVXXmHt2rXMnj2b559/Pn/5li1bGDFiRKF1169fz8svv3zGf
[…]
