+++ _sources/qmix.misc.rst.txt
@@ -1,0 +1,30 @@
+qmix.misc package
+=================
+
+Submodules
+----------
+
+qmix.misc.progbar module
+------------------------
+
+.. automodule:: qmix.misc.progbar
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+qmix.misc.terminal module
+-------------------------
+
+.. automodule:: qmix.misc.terminal
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: qmix.misc
+    :members:
+    :undoc-members:
+    :show-inheritance:

[…]
