### 2019-02-08 90d6e2f
Uncomment RespFnExponential
