### 2019-02-08 #1231 @garrettj403
Hi @FaustinCarter, I have added docs for all of the functions and classes [here](https://garrettj403.github.io/QMix/qmix.html). This includes an index [here](https://garrettj403.github.io/QMix/genindex.html) and a module index [here](https://garrettj403.github.io/QMix/py-modindex.html).
