### 2019-02-09 b17db1b
Add more info to testing comments
