+++ tests/test_harmonic_balance.py
@@ -1,6 +1,11 @@
 """Test the module that performs harmonic balance 
 (qmix.harmonic_balance.harmonic_balance).
 
+This module is relatively easy to test. Assuming that harmonic balance was
+performed correctly, the circuit should be 'balanced', meaning that the
+voltage drop across the junction should match the Thevenin equivalent circuit.
+See Section 4.3 in Garrett (2018) to understand what this means.
+
 """
 
 import numpy as np
@@ -17,45 +22,46 @@ NPTS = 101
 VBIAS = np.linspace(0, 2, NPTS)
 
 
-def test_relative_error_in_circuit():
+def test_relative_error_in_hb_solution():
     """ The harmonic_balance module can be tested by running the harmonic 
     balance process and then checking that the optimized voltage from the 
     harmonic_balance function does actually lead to a balanced circuit."""
 
     # Input parameters ------------------------------------------------------
 
-    NB = (15, 9)
-    NF = 2
-    NP = 2
-    N = NF * NP
+    NF = 2  # number of tones
+    NP = 2  # number of harmonics
+    N = NF * NP  # total number of signals
+    NB = (15, 9)  # number of Bessel functions to use
 
+    # Generate embedding circuit
     circuit = qmix.circuit.EmbeddingCircuit(NF, NP)
-
+    # Photon voltage
     circuit.vph[1] = 0.30
     circuit.vph[2] = 0.32
-
+    # Embedding voltage
     circuit.vt[1, 1] = circuit.vph[1] * 1.5
     circuit.vt[1, 2] = circuit.vph[1] * 0.1
     circuit.vt[2, 1] = circuit.vph[2] * 0.1
     circuit.vt[2, 2] = circuit.vph[2] * 0.01
-
+    # Embedding impedance
     circuit.zt[1, 1] = 0.3 - 1j*0.3
     circuit.zt[1, 2] = 0.3 - 1j*0.3
     circuit.zt[2, 1] = 0.3 - 1j*0.3
     circuit.zt[2, 2] = 0.3 - 1j*0.3
 
-
     # Run test ---------------------------------------------------------------
 
-    # Perform harmonic balance
-    vj, _, got_solution = harmonic_balance(circuit, RESP, NB, 
-                                                 stop_rerror=0.001, 
-                                                 mode='x')
+    # Perform harmonic balance to calculate voltage across junction (vj)
+    vj, _, solution_found = harmonic_balance(circuit, RESP, NB, 
+                                             stop_rerror=0.001, mode='x')
 
-    assert got_solution, "No solution found. Max iterations was reached!"
+    assert solution_found, "No solution found. Max iterations was reached!"
 
-    # This function will raise an exception if the relative error exceeds 
-    # the 'stop_error' value.
+    # This function will raise an exception if the error function exceeds 
+    # the 'stop_error' value. Here we will set this value to maximum
+    # error of 0.001. This is a relative error: error / vj. This is checked
+    # for every tone, harmonic and bias voltage.
     check_hb_error(vj, circuit, RESP, NB, stop_rerror= 0.001)
 
 
@@ -72,11 +78,11 @@ def test_error_handling():
     NP = 1
     N = NF * NP
 
+    # Embedding circuit
     circuit = qmix.circuit.EmbeddingCircuit(NF, NP)
     circuit.vph[1] = 0.30
     circuit.vt[1, 1] = circuit.vph[1] * 1.5
     circuit.zt[1, 1] = 0.3 - 1j*0.3
-
 
     # Run test ---------------------------------------------------------------
 
[…]
