### 2019-02-09 257b7b9
Fix rounding error in circuit test
