+++ environment-qmix.yaml
@@ -18,4 +18,4 @@ dependencies:
   - sphinx
   - sphinx_rtd_theme
   # for testing
-  - pytest
+  - pytest>=3.0
