### 2019-02-11 #1231 @PaulKGrimes
I have been able to confirm the operation of the experimental data analysis and impedance fitting example using my own experimental data.  This appears to work fairly well, and gives results as expected.  There could be some enhancements made to the software to make loading data from a variety of file types simpler - right now the data has to be reformatted to the style of data file used by John.

I have also been able to recreate some of the figures from my PhD thesis, calculating the mixer conversion gain and pumped IV curves for a 4 tone mixer (LO, USB, LSB and IF tones) which used a completely different multitone harmonic balance code, written by myself.  John and I use somewhat different definitions of mixer gain (particularly around the available power at the IF), but the calculated mixer gains are similar.  To me, this confirms that the software is operating broadly as expected, and accurately calculating the behavior of the SIS junction when used as a superheterodyne mixer.

However, the QMix code is extremely slow for 4 tones, taking approximately 5 hours to calculate the mixer currents using the standard response function in the multitone examplte.  The 3 tone version of the calculation (dropping the lower sideband) finishes in about 3 minutes.  I have some ideas about how this performance might be improved, but I don't think that that is necessary at this time since no claims are made about the performance of the software.

I think the only box that I'm not ready to tick is the Community Guidelines box.
