### 2019-02-11 0943237
Add CODE_OF_CONDUCT.md [ci skip]
