### 2019-02-13 bfd09dc
Add CODE_GUIDELINES.md [ci skip]
