### 2019-02-13 610a079
Add more info to CONTRIBUTING.md [ci skip]
