+++ CONTRIBUTING.md
@@ -1,28 +1,34 @@
-How to contribute to QMix 
-=========================
+Contributing to QMix 
+====================
 
-We welcome contributions to QMix via pull requests:
-1. fork the QMix repository to your own GitHub profile, 
-2. make any changes that you like, and then 
-3. create a pull request to merge your changes back into the main QMix repository. 
+We welcome contributions to QMix via the [GitHub issue tracker](https://github.com/garrettj403/QMix/issues) and via pull requests. Please see below for more information.
 
-I'll do my best to review your changes in a timely manner.
-
-**Note:** For information on how to download and install QMix, please see the [QMix website](https://garrettj403.github.io/QMix/setup.html).
+If you are new to GitHub, you can learn more about how to make contributions to GitHub projects [here](https://guides.github.com/introduction/flow/) and more about forking repositories [here](https://guides.github.com/activities/forking/).
 
 Bugs 
 ----
 
-If you find any bugs, please either create a new issue through the [GitHub Issue tracker](https://github.com/garrettj403/QMix/issues) with the ``bug`` label, or fix the issue yourself via a pull request (thank you in advance!)
+If you find any bugs in QMix, please either create a new issue through the [GitHub Issue tracker](https://github.com/garrettj403/QMix/issues) (with the ``bug`` label) or fix the issue yourself by forking the repository, making the necessary changes and then creating a pull request. Thank you in advance!
 
-New Features
-------------
+**Note:** Before reporting a bug, please ensure that you are using the latest version of QMix and that you are using the QMix virtual environment (``QMix/environment-qmix.yaml``).
 
-If you would like a new feature added, please create a new issue through the [GitHub Issue tracker](https://github.com/garrettj403/QMix/issues) with the ``enhancement`` label. I can likely advise how to implement the new feature, but there is no guarantee that the feature will be added.
+When reporting a bug, please include:
+   - your operating system,
+   - your Python version,
+   - your package versions (Numpy, SciPy and Matplotlib), and
+   - if applicable, a small code snippet that recreates the error.
 
-Problems Running or Installing QMix
------------------------------------
+Requesting New Features
+-----------------------
 
-If you have any problems getting QMix to work, first read through the [installation instructions](https://garrettj403.github.io/QMix/setup.html#installing-qmix-using-pip) online and then make sure you are using the [virtual environment](https://garrettj403.github.io/QMix/setup.html#configuring-python) (``QMix/environment.yml``) that I included with QMix.
+If you would like a new feature added to QMix, please create a new issue through the [GitHub Issue tracker](https://github.com/garrettj403/QMix/issues) (with the ``enhancement`` label). We can then discuss the improvements online, including whether the changes are necessary and how to implement the changes.
 
-If you are still having problems, please either [create a new issue](https://github.com/garrettj403/QMix/issues) or email me at garrettj403@gmail.com. I can also help you setup your simulations.
+Contributing New Code
+---------------------
+
+If you would like to contribute new code to the QMix project, please follow the [QMix code guidelines](https://github.com/garrettj403/QMix/blob/master/CODE_GUIDELINES.md).
+
+Code of Conduct
+---------------
+
+If you participate in the QMix project, you must follow our [Code of Conduct](https://github.com/garrettj403/QMix/blob/master/CODE_OF_CONDUCT.md).
