### 2019-02-13 66877f4
Fix minor issues in documentation [ci skip]
