### 2019-02-13 3800f29
Add info about license [ci skip]
