### 2019-02-13 3559e11
Minor tweaks to docs [ci skip]
