### 2019-02-13 2e17542
Add info for seeking support [ci skip]
