### 2019-02-13 4a137f9
Add info on seeking support and making contributions [ci skip]
