### 2019-02-13 cb3caff
Minor tweaks to docs [ci skip]
