### 2019-02-13 #1231 @garrettj403
Hi @PaulKGrimes, thank you for all your work validating the code. I am interested to hear about how to improve the performance of the 4 tone simulations, but we can talk about that later.

With respect to the Community Guidelines, I added more information to the [CONTRIBUTING.md](https://github.com/garrettj403/QMix/blob/master/CONTRIBUTING.md) file. I also added a [CODE_OF_CONDUCT.md](https://github.com/garrettj403/QMix/blob/master/CODE_OF_CONDUCT.md) (generated using GitHub's template) and some [CODE_GUIDELINES.md](https://github.com/garrettj403/QMix/blob/master/CODE_GUIDELINES.md).
