+++ paper/paper.bib
@@ -1,41 +1,27 @@
-
-@article{Tucker1979,
-author = {J. R. Tucker},
-doi = {10.1109/JQE.1979.1069931},
-journal = {IEEE Journal of Quantum Electronics},
-month = nov,
-number = 11,
-pages = {1234--1258},
-title = {{Quantum limited detection in tunnel junction mixers}},
-url = {http://ieeexplore.ieee.org/lpdocs/epic03/wrapper.htm?arnumber=1069931},
-volume = 15,
-year = 1979,
+@phdthesis{Garrett2018,
+author = {J. D. Garrett},
+school = {University of Oxford},
+title = {{A 230 GHz Focal Plane Array Using a Wide IF Bandwidth SIS Receiver}},
+type = {DPhil},
+year = 2018,
 }
 
-@article{Tucker1985,
-author = {J. R. Tucker and M. J. Feldman},
-doi = {10.1103/RevModPhys.57.1055},
-journal = {Reviews of Modern Physics},
-month = oct,
-number = 4,
-pages = {1055--1113},
-title = {{Quantum detection at millimeter wavelengths}},
-url = {https://link.aps.org/doi/10.1103/RevModPhys.57.1055},
-volume = 57,
-year = 1985,
+@article{Garrett2019,
+  author       = {J. D. Garrett and G. Yassin},
+  title        = {{garrettj403/QMix  (Version v1.0.1)}},
+  month        = feb,
+  year         = 2019,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.2538163},
+  url          = {http://doi.org/10.5281/zenodo.2557839}
 }
 
-@article{Withington2003,
-author = {S. Withington and P. Kittara and G. Yassin},
-doi = {10.1063/1.1576515},
-journal = {Journal of Applied Physics},
-month = jun,
-number = 12,
-pages = {9812--9822},
-title = {{Multitone quantum simulations of saturating tunnel junction mixers}},
-url = {http://aip.scitation.org/doi/10.1063/1.1576515},
-volume = 93,
-year = 2003,
+@phdthesis{Kittara2002,
+author = {P. Kittara},
+school = {University of Cambridge},
+title = {{The Development of a 700 GHz SIS Mixer with Nb Finline Devices: Nonlinear Mixer Theory, Design Techniques and Experimental Investigation}},
+type = {DPhil},
+year = 2002,
 }
 
 @article{Kittara2007,
@@ -51,32 +37,6 @@ volume = 101,
 year = 2007,
 }
 
-@phdthesis{Kittara2002,
-author = {P. Kittara},
-school = {University of Cambridge},
-title = {{The Development of a 700 GHz SIS Mixer with Nb Finline Devices: Nonlinear Mixer Theory, Design Techniques and Experimental Investigation}},
-type = {DPhil},
-year = 2002,
-}
-
-@phdthesis{Garrett2018,
-author = {J. D. Garrett},
-school = {University of Oxford},
-title = {{A 230 GHz Focal Plane Array Using a Wide IF Bandwidth SIS Receiver}},
-type = {DPhil},
-year = 2018,
-}
-
-@article{Garrett2019,
-  author       = {J. D. Garrett and G. Yassin},
-  title        = {{garrettj403/QMix  (Version v1.0.0)}},
-  month        = jan,
-  year         = 2019,
-  publisher    = {Zenodo},
-  doi          = {10.5281/zenodo.2538163},
-  url          = {https://doi.org/10.5281/zenodo.2538163}
-}
-
 @article{Werthamer1966,
 author = {N. R. Werthamer},
 doi = {10.1103/PhysRev.147.255},
[…]
