+++ qmix/exp/exp_data.py
@@ -32,7 +32,7 @@ from scipy.interpolate import UnivariateSpline
 import qmix
 from qmix.exp.if_data import dcif_data, if_data
 from qmix.exp.iv_data import dciv_curve, iv_curve
-from qmix.exp.parameters import file_structure, params
+from qmix.exp.parameters import params
 from qmix.harmonic_balance import harmonic_balance
 from qmix.mathfn.filters import gauss_conv
 from qmix.mathfn.misc import slope_span_n
@@ -59,6 +59,20 @@ _plot_params = {'dpi': 500, 'bbox_inches': 'tight'}
 # by using:  "# pragma: no cover"
 
 # TODO: Fix docstrings (esp. keyword args)
+
+
+# FILE HIERARCHY --------------------------------------------------------------
+
+"""Default file hierarchy to use when plotting experimental data."""
+_file_structure = {'DC IV data':          '01_dciv/',
+                   'Pumped IV data':      '02_iv_curves/',
+                   'IF data':             '03_if_data/',
+                   'Impedance recovery':  '04_impedance/',
+                   'IF noise':            '05_if_noise/',
+                   'Noise temperature':   '06_noise_temp/',
+                   'IF spectrum':         '07_spectrum/',
+                   'Overall performance': '08_overall_performance/',
+                   'CSV data':            '09_csv_data/'}
 
 
 # CLASSES FOR RAW DATA -------------------------------------------------------
@@ -388,21 +402,48 @@ class RawData0(object):
         else:
             fig1.savefig(fig_name, **_plot_params)
 
-    def plot_all(self, fig_folder, **kw):  # pragma: no cover
-        """Plot all DC data.
+    def plot_all(self, fig_folder, sub_folder=None, **kw):  # pragma: no cover
+        """Plot all DC data and save it to a specified directory.
 
-        This function will save the data in the standard format.
+        This method will call the following methods: ``plot_dciv``, 
+        ``plot_offset``, ``plot_rdyn`` and ``plot_if_noise``.
+        
+        These figures will be put in ``fig_folder/sub_folder``. Note that if
+        ``sub_folder`` is left as ``None``, this method will use the default 
+        file structure (see ``qmix.exp.exp_data._file_structure``). If you 
+        would instead like the figures to go into ``fig_folder``, set this 
+        argument as an empty string ("").
 
         Args:
-            fig_folder: directory where the figures go
-            kw (dict): keyword arguments (not used)
+            fig_folder (str): directory where the figures go
+            sub_folder (str): sub-directory where the DC I-V figures go
+            kw: keyword arguments that will be passed to the plotting methods
 
         """
 
-        self.plot_dciv(fig_folder     + '01_dciv/dciv.png', **kw)
-        self.plot_offset(fig_folder   + '01_dciv/dciv-offset.png', **kw)
-        self.plot_rdyn(fig_folder     + '01_dciv/dciv-rdyn.png', **kw)
-        self.plot_if_noise(fig_folder + '01_dciv/dcif-shot-noise.png', **kw)
+        # Folder for DC data
+        if sub_folder is None:
+            sub_folder = _file_structure['DC IV data']
+            folder = os.path.join(fig_folder, sub_folder)
+        else:
+            folder = os.path.join(fig_folder, sub_folder)
[…]
