+++ qmix/exp/__init__.py
@@ -2,5 +2,6 @@
 experimental data. """
 
 from qmix.exp.exp_data import (RawData, RawData0,
-                               plot_overall_results, plot_if_spectrum,
-                               check_iv_if_matching, generate_filename_from_f)
+                               plot_overall_results, plot_if_spectrum)
+#, check_iv_if_matching)
+#, generate_filename_from_f)
