### 2019-02-27 090631b
Change plotting functions to accept and return Matplotlib axes [ci skip]
