+++ qmix/exp/exp_data.py
@@ -78,67 +78,69 @@ _file_structure = {'DC IV data':          '01_dciv/',
 # CLASSES FOR RAW DATA -------------------------------------------------------
 
 class RawData0(object):
-    """Class for experimental DC data (i.e., no LO present).
+    """Class for importing and analyzing experimental DC data (with no LO).
 
     Note:
 
         See ``qmix.exp.parameters.params`` for all possible keyword arguments.
 
     Args:
-        dciv_file: file path for unpumped I-V data
+        dciv_file: file path for unpumped I-V curve
 
     Keyword arguments:
-        dcif_file (str): file path for unpumped IF data
+        dcif_file (str): file path for unpumped IF data (default is None)
         area (float): area of the junction in um^2 (default is 1.5)
         comment (str): add comment to this instance (default is '')
-        filter_data (bool): smooth/filter the I-V data (default is True)
-        i_fmt (str): units for current ('uA', 'mA', etc.)
-        igap_guess (float): Gap current estimate (used to temporarily normalize the input data during filtering)
-        ioffset (float): current offset
-        rseries (float): Remove the effect of a series resistance
-        v_fmt (str): units for voltage ('mV', 'V', etc.)
-        v_smear (float): smear DC I-V by this amount when generating response function
-        vgap_guess (float): Gap voltage estimate (used to temporarily normalize the input data during filtering)
-        voffset (float): voltage offset
-        vmax (float): maximum voltage value to import (in case the I-V curve saturates at some point)
-        vrsg (float): The voltage to calculate the subgap resistance at
+        filter_data (bool): smooth the I-V data (default is True)
+        i_fmt (str): units for current: 'uA', 'mA', etc. (default is 'mA')
+        igap_guess (float): Gap current estimate, used to temporarily normalize the input data while filtering (default is 2e-4 A)
+        ioffset (float): current offset (default is None)
+        rseries (float): Remove influence of a series resistance (default is None)
+        v_fmt (str): units for voltage: 'mV', 'V', etc. (default is 'mV')
+        v_smear (float): smear DC I-V by this amount when generating smeared response function (default is 0.020)
+        vgap_guess (float): Gap voltage estimate, used to temporarily normalize the input data while filtering (default is 2.7e-3 V)
+        voffset (float): voltage offset (default is None)
+        vmax (float): maximum voltage value to import, used in case the I-V curve saturates at some point
+        vrsg (float): The voltage at which to calculate the subgap resistance (default is 2e-3 V)
 
     """
 
-    def __init__(self, dciv_file, dcif_file=None, **kwargs):
+    def __init__(self, dciv_file, dcif_file=None, **kw):
 
         # Import keyword arguments
         tmp = deepcopy(params)
-        tmp.update(kwargs)
-        kwargs = tmp
+        tmp.update(kw)
+        kw = tmp
 
         # Unpack keyword arguments
-        comment = kwargs['comment']
-        v_smear = kwargs['v_smear']
-        vleak = kwargs['vleak']
-        area = kwargs['area']
-        verbose = kwargs['verbose']
+        comment = kw['comment']
+        v_smear = kw['v_smear']
+        vleak   = kw['vleak']
+        area    = kw['area']
+        verbose = kw['verbose']
 
-        self.kwargs = kwargs
+        self.kwargs    = kw
[…]
