### 2019-02-27 27e3d2e
Fix documentation for experimental parameters [ci skip]
