### 2019-02-28 7e47a70
Make plotting function accept and return matplotlib axis [ci skip]
