### 2019-02-28 bb7f192
Describe all class attributes in detail [ci skip]
