+++ qmix/circuit.py
@@ -1,21 +1,24 @@
 """ This module contains classes and functions to describe the embedding 
 circuit.
 
-**Description**
-
+**Description:**
+    
     In experimental systems, SIS junctions are embedded within complex RF 
     networks. These networks are referred to as the embedding circuit. Since
     all of the components in these embedding circuits are **linear**, the 
     embedding circuit can be reduced to a Thevenin equivalent circuit for 
     **each tone and harmonic.**
-
+    
     To fully describe the embedding circuit, 3 bits of information are needed
     for each signal that is applied to the junction:
-
+    
         1. the frequency of the applied signal,
         2. the Thevenin voltage of the embedding circuit at this freq., and
         3. the Thevenin impedance of the embedding circuit at this freq.
 
+    The main class in this module (``EmbeddingCircuit``) allows the user to 
+    build an embedding circuit in the proper format.
+    
 """
 
 import re
@@ -27,67 +30,124 @@ from qmix.misc.terminal import cprint
 
 # EMBEDDING CIRCUIT CLASS ----------------------------------------------------
 
-
 class EmbeddingCircuit(object):
-    """Class for describing the embedding circuit. 
+    """Class for building and describing the embedding circuit. 
 
-    This includes the voltages, impedances and frequencies of all signals 
-    applied to the junction.
+    This includes the frequencies, Thevenin voltages and Thevenin impedances 
+    of all signals applied to the junction.
 
-    Creating an instance of this class will set the sizes and data types
-    of all of the class attributes, but the actual values will need to be set
-    manually. In this way, this class is sort of like a struct. The
-    class attributes that have to be set manually are:
-
-       - ``vph``:  photon voltage (freq / gap freq)
-       - ``vt``:  Thevenin voltage (normalized to the gap voltage)
-       - ``zt``:  Thevenin impedance (normalized to the normal resistance)
-
-    Note:
-
-        1. Unless specified otherwise, **all input values are normalized**. 
-        The voltages are normalized to the gap voltage (Vgap), the 
-        resistances are normalized to the normal-state resistance (Rn), the 
-        currents are normalized to Igap=Vgap/Rn, and the frequencies are 
-        normalized to the gap frequency fgap=e*Vgap/h.
-
-        2. The photon voltage is defined as hf/e. Therefore, the normalized 
-        photon voltage is equal to the frequency of the input signal divided
-        by the the gap frequency. **Normalized photon voltage is equal to the
-        normalized frequency.**
+    Note: 
+    
+        Unless specified otherwise, **all input values are normalized**. The 
+        voltages are normalized to the gap voltage, resistances are normalized 
+        to the normal-state resistance, currents are normalized to gap current, 
+        and frequencies are normalized to the gap frequency. Refer to the 
+        argument descriptions below to see if the value is normalized or not.
+        
+        Creating an instance of this class will set the sizes and data types
+        of all of the class attributes, but the actual values will need to be 
+        set manually. In this way, this class is sort of like a fancy struct. 
+        The class attributes that have to be set manually are:
+    
+           - ``vph``:  photon voltage (freq / gap freq)
+           - ``vt``:  Thevenin voltage (normalized to the gap voltage)
+           - ``zt``:  Thevenin impedance (normalized to the normal resistance)
+    
+        The photon voltage is defined as hf/e. Therefore, the normalized photon
+        voltage is equal to the fundamental frequency divided by the the gap 
+        frequency (i.e., the normalized frequency). 
 
     Example:
-
+    
         To create an instance of the embedding circuit class with 2 tones and 
         3 harmonices, you would set:
-
             ``cct = EmbeddingCircuit(2, 3)``
[…]
