### 2019-02-28 4facce8
Describe Resp. Fn. types, keyword args. and attributes in detail [ci skip]
