+++ qmix/respfn.py
@@ -3,40 +3,83 @@ SIS junction.
 
 Upon initialization, these classes will:
 
-   1. Load or calculate the DC I-V curve (imagainary part of resp. func.)
+   1. Load or calculate the DC I-V curve (which is the imagainary part of 
+      response function).
 
       - The I-V curve can either be loaded from a csv file or generated from
-        an I-V curve model (e.g., polynomial model, exponential model, etc.),
+        an I-V curve model (e.g., polynomial model, exponential model, etc.).
 
-   2. Calculate the Kramers-Kronig transform of the DC I-V curve
+   2. Calculate the Kramers-Kronig transform of the DC I-V curve (which is the
+      real part of the response function).
 
-   3. Setup the density of data points to optimize interpolation
+   3. Setup the density of the data points to optimize interpolation.
 
-      - The I-V curve needs to be interpolated and the number of points has a
-        large affect on the speed of this, but at the same time there needs to
-        be enough points to accurately represent the original I-V curve.
+      - The response function needs enough data points that it can be 
+        interpolated accurately, but at the same time, not so many points 
+        that the interpolation takes too long.
 
-   4. Calculate cubic spline fits for the I-V curve and the KK transform
+   4. Calculate cubic spline fits for the I-V curve and the KK transform.
 
       - Doing this once at the start allows the data to be interpolated very
         quickly later on.
 
    5. Determine the derivative of the I-V curve and the KK transform based on
-      the spline fits
+      the spline fits.
 
-Then when needed, the classes allow the user to interpolated the DC I-V curve,
+      - These derivatives aren't needed by QMix, but they are needed for the
+        equations from Tucker theory (Tucker and Feldman, 1985).
+
+Once initialized, the classes allow the user to interpolated the DC I-V curve,
 the KK transform, the derivative of the I-V curve, the derivative of the KK
-transform, or the response function (a complex value). This is set up to run
-very quickly.
+transform, and the response function (a complex value). These classes are 
+optimized to interpolate very quickly.
+
+There are several different response functions classes contained within this
+module:
+
+   - Response functions generated from voltage and current data:
+   
+      - ``RespFn``: This is the base class for all of the other response 
+        function classes. This class will generate a response function 
+        based on a DC I-V curve (i.e., DC voltage and current data). Note that 
+        this class assumes that you have "pre-processed" the data. This means 
+        that it will use the voltage and current data to generate the 
+        interpolation directly. Normally, you want to have more data points 
+        around curvier regions in order to minimize how much time the 
+        interpolation takes. If you haven't done this, it is a good idea to use
+        ``RespFnFromIVData`` instead.
+
+      - ``RespFnFromIVData``: This class will generate a response function 
+        based on a DC I-V curve (i.e., DC voltage and current data). Unlike
+        ``RespFn``, this class will resample the response function in order to
+        optimize the interpolation.
+
+   - Response functions generated from mathematical models:
+
+      - ``RespFnPerfect``: This class will generate a response function based
+        on an ideal DC I-V curve (i.e., the subgap current is exactly zero
+        below the gap voltage and, above the gap, it is exactly equal to the
+        bias voltage, assuming normalized values). This DC I-V curve has an
+        infinitely sharp transition.
+
+      - ``RespFnPolynomial``: This class will generate a response function 
[…]
