### 2019-02-28 cb10e3f
Get fig object from axis if needed [ci skip]
