+++ qmix/exp/exp_data.py
@@ -254,6 +254,8 @@ class RawData0(object):
         # Plot DC I-V curve
         if ax is None:
             fig, ax = plt.subplots()
+        else:
+            fig = ax.get_figure()
         ax.plot(v_mv, i_ua, label=lgd_str1)
         # Label gap
         ax.plot(self.vgap * 1e3, i_at_gap, 
@@ -279,6 +281,7 @@ class RawData0(object):
         if fig_name is not None:
             fig.savefig(fig_name, **_plot_params)
             plt.close(fig)
+            return
         else:
             return ax
 
@@ -311,6 +314,8 @@ class RawData0(object):
         # Plot offset (to make sure it was corrected properly)
         if ax is None:
             fig, ax = plt.subplots()
+        else:
+            fig = ax.get_figure()
         ax.plot(v_mv[mask], i_ua[mask])
         ax.set_xlim([-0.2, 0.2])
         ax.set_xlabel(r'Bias Voltage (mV)')
@@ -320,6 +325,7 @@ class RawData0(object):
         if fig_name is not None:
             fig.savefig(fig_name, **_plot_params)
             plt.close(fig)
+            return
         else:
             return ax
 
@@ -354,6 +360,8 @@ class RawData0(object):
         # Plot dynamic resistance
         if ax is None:
             fig, ax = plt.subplots()
+        else:
+            fig = ax.get_figure()
         ax.semilogy(v_mv, 1 / rdyn)
         ax.set_xlabel(r'Bias Voltage (mV)')
         ax.set_ylabel(r'Dynamic Resistance ($\Omega$)')
@@ -362,6 +370,7 @@ class RawData0(object):
         if fig_name is not None:
             fig.savefig(fig_name, **_plot_params)
             plt.close(fig)
+            return
         else:
             return ax
 
@@ -399,6 +408,8 @@ class RawData0(object):
         # Plot static resistance
         if ax is None:
             fig, ax = plt.subplots()
+        else:
+            fig = ax.get_figure()
         ax.plot(v_mv, r_stat)
         ax.set_xlabel(r'Bias Voltage (mV)')
         ax.set_ylabel(r'Static Resistance ($\Omega$)')
@@ -408,6 +419,7 @@ class RawData0(object):
         if fig_name is not None:
             fig.savefig(fig_name, **_plot_params)
             plt.close(fig)
+            return
         else:
             return ax
 
@@ -450,6 +462,7 @@ class RawData0(object):
             fig, (ax1, ax2) = plt.subplots(2, sharex=True, figsize=(6, 9))
         else:
             ax1, ax2 = ax
+            fig = ax1.get_figure()
         plt.subplots_adjust(hspace=0., wspace=0.)
 
         # Plot DC I-V curve
@@ -480,6 +493,7 @@ class RawData0(object):
         if fig_name is not None:
             fig.savefig(fig_name, **_plot_params)
             plt.close(fig)
+            return
         else:
             return ax1, ax2
 
@@ -797,6 +811,8 @@ class RawData(object):
 
         if ax is None:
             fig, ax = plt.subplots()
+        else:
+            fig = ax.get_figure()
         ax.plot(dciv.voltage * vmv, dciv.current * iua, label="Unpumped")
         ax.plot(self.voltage * vmv, self.current * iua, 'r', label="Pumped")
         ax.set_xlabel(r'Bias Voltage (mV)')
@@ -809,6 +825,7 @@ class RawData(object):
         if fig_name is not None:
             fig.savefig(fig_name, **_plot_params)
             plt.close(fig)
+            return
         else:
             return ax
 
@@ -835,6 +852,8 @@ class RawData(object):
[…]
