### 2019-02-28 c7c4859
Minor fixes
