### 2019-03-01 4016051
Add use cases to method docstrings, and add more frequency/power units
