+++ qmix/circuit.py
@@ -6,7 +6,7 @@ circuit.
     In experimental systems, SIS junctions are embedded within complex RF 
     networks. These networks are referred to as the embedding circuit. Since
     all of the components in these embedding circuits are **linear**, the 
-    embedding circuit can be reduced to a Thevenin equivalent circuit for 
+    embedding circuit can be reduced to a **Thevenin equivalent circuit** for
     **each tone and harmonic.**
     
     To fully describe the embedding circuit, 3 bits of information are needed
@@ -24,11 +24,14 @@ circuit.
 import re
 
 import numpy as np
+import scipy.constants as sc
 
 from qmix.misc.terminal import cprint
 
 
 # EMBEDDING CIRCUIT CLASS ----------------------------------------------------
+
+# TODO: Use @property decorators to avoid setters/getters
 
 class EmbeddingCircuit(object):
     """Class for building and describing the embedding circuit. 
@@ -36,113 +39,163 @@ class EmbeddingCircuit(object):
     This includes the frequencies, Thevenin voltages and Thevenin impedances 
     of all signals applied to the junction.
 
-    Note: 
-    
-        Unless specified otherwise, **all input values are normalized**. The 
-        voltages are normalized to the gap voltage, resistances are normalized 
-        to the normal-state resistance, currents are normalized to gap current, 
-        and frequencies are normalized to the gap frequency. Refer to the 
-        argument descriptions below to see if the value is normalized or not.
-        
+    Note:
+
+        Unless specified otherwise, **all input values are normalized**. The
+        voltages are normalized to the gap voltage, resistances are normalized
+        to the normal-state resistance, currents are normalized to gap current,
+        and frequencies are normalized to the gap frequency. Refer to the
+        argument descriptions below to see if the value should be normalized or
+        not.
+
         Creating an instance of this class will set the sizes and data types
-        of all of the class attributes, but the actual values will need to be 
-        set manually. In this way, this class is sort of like a fancy struct. 
+        of all of the class attributes, but the actual values will need to be
+        set manually. In this way, this class is sort of like a fancy struct.
         The class attributes that have to be set manually are:
-    
+
            - ``vph``:  photon voltage (freq / gap freq)
            - ``vt``:  Thevenin voltage (normalized to the gap voltage)
            - ``zt``:  Thevenin impedance (normalized to the normal resistance)
-    
-        The photon voltage is defined as hf/e. Therefore, the normalized photon
-        voltage is equal to the fundamental frequency divided by the the gap 
-        frequency (i.e., the normalized frequency). 
+
+        See the attribute descriptions below for information about how to do
+        this.
 
     Example:
-    
-        To create an instance of the embedding circuit class with 2 tones and 
-        3 harmonices, you would set:
-            ``cct = EmbeddingCircuit(2, 3)``
-        Then, to set the Thevenin voltage of the 3rd harmonic of the 2nd tone
-        to 0.3, you would set:
-            ``cct.vt[2,3] = 0.3``
-        For each signal, you then have to repeat this to set each voltage, 
-        impedance and photon voltage.
-        
-    Keyword Args:
-        num_f (int, default is 1): Number of fundamental frequencies (tones) 
+
+        To create an instance of the embedding circuit class with 2 tones and
+        3 harmonices:
+
+        >>> cct1 = EmbeddingCircuit(2, 3, name='Test 1')
+        >>> print(cct1)
+        Embedding circuit (Tones:2, Harmonics:3): Test 1
+
+        Once initialized, you can begin defining the properties of the
+        embedding circuit. I normally start with the photon voltages (or,
+        equivalently, the normalized frequencies):
+
+        >>> cct1.vph[1] = 0.30  # first tone
[…]
