### 2019-03-01 a751b9d
Fix error in docstring [ci skip]
