### 2019-03-01 8c89d28
Add use cases to docstrings, and use methods for interpolation, not attributes
