### 2019-03-01 2bfe357
Drop Python 3.5-dev and 3.6-dev
