+++ tests/README.md
@@ -9,32 +9,26 @@ This directory contains tests for all of the modules within the ``QMix`` package
 
 The tests for these packages are the most important. ``qmix.circuit``, ``qmix.respfn`` and ``qmix.harmonic_balance`` are all very easy to test. ``qmix.qtcurrent``, on the other hand, is very difficult to test because it tests how the quasiparticle tunneling currets are calculated.
 
-Note: 
+**Note:** The ``qmix.exp.exp_data`` module is for analyzing experimental data. This module does not effect QMix simulations.
 
-    The ``qmix.exp.exp_data`` module is for analyzing experimental data. This module does not effect QMix simulations.
-
-Automatic Testing
+Automated Testing
 -----------------
 
 [![Build Status](https://travis-ci.org/garrettj403/QMix.svg?branch=master)](https://travis-ci.org/garrettj403/QMix)
 
 Everytime there is a new commit, ``QMix`` is automatically tested using [Travis CI](https://travis-ci.org/garrettj403/QMix). 
 
-Note: 
-    
-    If you make some changes that do not affect the code, you can prevent Travis CI from running by including ``[ci skip]`` at the end of the commit message.
+**Note:** If you make some changes that do not affect the code, you can prevent Travis CI from running by including ``[ci skip]`` at the end of the commit message.
 
 Testing QMix on Your Machine
 ----------------------------
 
 From the ``QMix/`` directory, run:
-
 ```bash
 pytest --verbose --color=yes tests/
 ```
 
 To test coverage, run:
-
 ```bash
 pytest --verbose --color=yes --cov=qmix/ tests/
 ```
@@ -48,10 +42,8 @@ then open the report that is generated in ``htmlcov/index.html``.
 Testing Individual Modules
 --------------------------
 
-As an example, from the ``QMix/tests/`` directory you can run:
-
+As an example, from the ``QMix/tests/`` directory, you can run:
 ```bash
 pytest test_harmonic_balance.py
 ```
-
 to test the harmonic balance module
