+++ examples/eg-230-data/README.md
@@ -1,12 +1,11 @@
 Example Experimental Data
 =========================
 
-This directory contains example experimental data that can be used for running the example workflows. This data was taken from a 230 GHz SIS mixer (see my thesis). Do not use this data for anything else besides testing the workflows.
+This directory contains example experimental data that can be used for running the example workflows. This data was taken from a 230 GHz SIS mixer (Garrett, 2018). Do not use this data for anything else besides testing the workflows.
 
-This includes:
-   1. ``dciv.csv``: DC I-V data
-   2. ``dcif.csv``: DC IF data
-   3. ``f230.2_iv.csv``: pumped I-V curve at 230.2 GHz
-   4. ``f230.2_if-hot.csv``: IF output data with hot load at 230.2 GHz
-   5. ``f230.2_if-cold.csv``: IF output data with cold load at 230.2 GHz
-
+These files include:
+   - ``dciv.csv``: DC I-V data (with no LO pumping)
+   - ``dcif.csv``: DC IF power data (with no LO pumping)
+   - ``f230.2_iv.csv``: pumped I-V curve at 230.2 GHz
+   - ``f230.2_if-hot.csv``: IF output data from hot load at 230.2 GHz
+   - ``f230.2_if-cold.csv``: IF output data from cold load at 230.2 GHz
