+++ examples/README.md
@@ -4,7 +4,6 @@ Example Workflows
 This directory contains 3 example workflows:
    1. ``single-tone-simple.ipynb``: a simple one tone simulation of an SIS junction
    2. ``multi-tone.ipynb``: a multi-tone simulation (heterodyne mixing) of an SIS junction
-   3. ``analyze-experimental-data.ipynb``: analyze experimental data from an SIS mixer
+   3. ``analyze-experimental-data.ipynb``: analyzing experimental data from an SIS mixer
 
 This directory contains some example experimental data, which is used to run these examples. This data is contained within ``eg-230-data/``.
-
