+++ tests/exp-data/README.md
@@ -1,11 +1,11 @@
 Example Experimental Data
 =========================
 
-This directory contains example experimental data that can be used to test the ``QMix`` package. This data was taken from a 230 GHz SIS mixer (see my thesis). Do not use this data for anything else besides testing the ``QMix`` package.
+This directory contains example experimental data that can be used to test the ``QMix`` package. This data was taken from a 230 GHz SIS mixer (Garrett, 2018). Do not use this data for anything else besides testing the ``QMix`` package.
 
-This includes:
-   1. ``dciv-data.csv``: measured DC I-V data (no LO)
-   2. ``dcif-data.csv``: measured IF power (no LO)
-   3. ``f230_0_iv.csv``: measured pumped I-V curve at 230 GHz
-   4. ``f230_0_if-hot.csv``: measured IF output power from a hot black body load at 230 GHz
-   5. ``f230_0_if-cold.csv``: measured IF output data from a hot black body load at 230 GHz
+These files include:
+   - ``dciv-data.csv``: measured I-V data with no LO present
+   - ``dcif-data.csv``: measured IF power with no LO present
+   - ``f230_0_iv.csv``: measured I-V data with LO pumping at 230 GHz
+   - ``f230_0_if-hot.csv``: measured IF output power from a hot black body load (room temperature Eccosorb) at 230 GHz
+   - ``f230_0_if-cold.csv``: measured IF output data from a hot black body load (Eccosorb immersed in liquid nitrogen) at 230 GHz
