### 2019-03-01 fe83c78
Remove old functions from import (previously commented out) [ci skip]
