### 2019-03-01 192d3d3
Update examples with new plotting functionality [ci skip]
