### 2019-03-01 e0144d1
Add ability to interpolate derivatives
