### 2019-03-02 6ae3051
Fix docstrings for DCIVData and DCIFData [ci skip]
