+++ _sources/analyze-experimental-data.ipynb.txt
@@ -6,9 +6,9 @@
    "source": [
     "# Example #3: Analyzing experimental data\n",
     "\n",
-    "This notebook will show a brief example of how QMix can be used to analyze experimental data. Note that a wide range of parameters can be passed to these functions in order to import data that has been stored in different formats.\n",
+    "This notebook will show a brief example of how QMix can be used to analyze experimental data. Note that a wide range of parameters can be passed to these functions in order to import data that has been stored in different formats. Please see ``params`` in ``qmix/exp/parameters.py`` for a full list.\n",
     "\n",
-    "For the purposes of this notebook, example data has been included in a directory called ``eg-230-data/``."
+    "For the purposes of this notebook, example data has been included in a directory called ``eg-230-data/``. This is experimental data that was taken from an SIS device illuminated by a local-oscillator (LO) at 230 GHz."
    ]
   },
   {
@@ -29,8 +29,9 @@
    "metadata": {},
    "source": [
     "There are two main classes that can be used to analyze experimental data:\n",
-    "- ``qmix.exp.RawData0`` which is used for DC data (no LO pumping), and\n",
-    "- ``qmix.exp.RawData`` which is used to analyze pumped data."
+    "\n",
+    "   - ``qmix.exp.RawData0``, which is used for DC data (no LO pumping), and\n",
+    "   - ``qmix.exp.RawData``, which is used to analyze pumped data (with LO illumination)."
    ]
   },
   {
@@ -39,15 +40,80 @@
    "source": [
     "## Import DC I-V data\n",
     "\n",
-    "This class will automatically load, filter, and analyze the DC I-V data. It will also automatically correct any offsets in the current or voltage data, and calculate the IF noise.\n",
+    "The DC experimental data is located in two different files:\n",
     "\n",
-    "- ``dciv.csv`` is the DC current vs. bias voltage (no LO pumping)\n",
-    "- ``dcif.csv`` is the IF power vs. bias voltage (no LO pumping)"
+    "   - ``dciv.csv``: DC current vs. bias voltage (no LO pumping)\n",
+    "   - ``dcif.csv``: IF power vs. bias voltage (no LO pumping)\n",
+    "   \n",
+    "Both files are CSV files with one column for voltage and one column for current or IF power."
    ]
   },
   {
    "cell_type": "code",
    "execution_count": 2,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "DC Bias Voltage (mV), DC Tunneling Current (mA)\r\n",
+      "0.016102, 0.007249\r\n",
+      "0.017391, 0.007265\r\n",
+      "0.019324, 0.007361\r\n",
+      "0.021257, 0.007378\r\n",
+      "0.023512, 0.007426\r\n",
+      "0.026090, 0.007474\r\n",
+      "0.028345, 0.007587\r\n",
+      "0.030600, 0.007619\r\n",
+      "0.032856, 0.007668\r\n"
+     ]
+    }
+   ],
+   "source": [
+    "%%bash\n",
+    "head eg-230-data/dciv.csv"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 3,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "DC Bias Voltage (mV), IF Power (A.U.)\r\n",
+      "0.015780, 0.009150\r\n",
[…]
