+++ _sources/multi-tone.ipynb.txt
@@ -6,17 +6,16 @@
    "source": [
     "# Example #2: Simulating an SIS mixer\n",
     "\n",
-    "- This is a multi-tone simultion (i.e., heterodyne mixing)\n",
-    "    - The input consists of 3 signals\n",
-    "        - a strong local-osciallator at 230 GHz\n",
-    "        - a weak RF signal at 232 GHz \n",
-    "        - an IF output at 2 GHz\n",
-    "    - The embedding circuit is included\n",
-    "    - The response function is generated from an experimental I-V curve\n",
-    "\n",
-    "- This simulation calculates:\n",
-    "    - The pumped I-V curve\n",
-    "    - The mixer's gain"
+    "- This example will simulate a simple SIS mixer.\n",
+    "- The input will consists of 3 signals:\n",
+    "    - a strong local-osciallator (LO) signal at 230 GHz,\n",
+    "    - a weak RF signal at 232 GHz, and\n",
+    "    - an IF output at 2 GHz.\n",
+    "- An arbitrary embedding circuit is included (therefore, harmonic balance is required).\n",
+    "- The response function is generated from an experimental DC I-V curve.\n",
+    "- This simulation will calculate:\n",
+    "    - the pumped I-V curve, and\n",
+    "    - the mixer's gain."
    ]
   },
   {
@@ -32,8 +31,6 @@
     "import scipy.constants as sc \n",
     "import matplotlib.pyplot as plt\n",
     "\n",
-    "import qmix.misc.terminal\n",
-    "from qmix.misc.terminal import cprint\n",
     "from qmix.mathfn.misc import slope"
    ]
   },
@@ -41,8 +38,8 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "## Define junction properties\n",
-    "- These are common parameters that describe the electrical properties of SIS junctions"
+    "## Define the junction properties\n",
[…]
