+++ docs/source/index.rst
@@ -1,15 +1,9 @@
-.. mixer documentation master file, created by
-   sphinx-quickstart on Thu Aug  4 13:43:26 2016.
-   You can adapt this file completely to your 
-   liking, but it should at least contain the 
-   root `toctree` directive.
-
 QMix: Quantum Mixing Software
 =============================
 
-QMix is a software package for simulating the quasiparticle tunneling currents in Superconductor/Insulator/Superconductor (SIS) junctions. These junctions are commonly used for heterodyne mixing at millimeter and submillimeter wavelengths. QMix can be used to simulate the performance of SIS mixers, investigate their experimental results and optimize their operation.
+QMix is a software package for simulating the quasiparticle tunneling currents in Superconductor/Insulator/Superconductor (SIS) junctions. In radio astronomy, these junctions are used for heterodyne mixing at millimeter and submillimeter wavelengths. QMix can be used to simulate their performance, investigate their experimental results and optimize their operation.
 
-In order to calculate the quasiparticle tunneling currents, QMix uses *multi-tone spectral domain analysis* (see references for more information). This makes QMix ideal for simulating higher-order harmonics, power saturation, sub-harmonic pumping and frequency multiplication. 
+In order to calculate the quasiparticle tunneling currents, QMix uses *multi-tone spectral domain analysis* (see :ref:`references` for more information). Among other applications, this makes QMix ideal for simulating power saturation, higher-order harmonics, sub-harmonic pumping and frequency multiplication.
 
 **Repository address:** https://github.com/garrettj403/QMix/
 
@@ -17,8 +11,8 @@ Contents
 ========
 
 .. toctree::
-   :maxdepth: 2
-   :numbered:
+   :maxdepth: 1
+   :titlesonly:
    
    setup
    qmix
