### 2019-03-06 a550da0
Set up for coveralls
