+++ .travis.yml
@@ -12,3 +12,6 @@ install:
   - pip install .
 # command to run tests
 script: pytest
+# for coveralls.io
+after_success:
+  - coveralls
