+++ _sources/index.rst.txt
@@ -11,8 +11,8 @@ Contents
 ========
 
 .. toctree::
-   :maxdepth: 1
-   :titlesonly:
+   :maxdepth: 3
+   :numbered:
    
    setup
    qmix
