+++ qmix/exp/parameters.py
@@ -1,70 +1,153 @@
-""" This module contains a dictionary of parameters that are used by 
-qmix.exp.exp_data.py to load and analyze experimental data.
+""" This module contains a dictionary of parameters (i.e., ``params``) that is
+used by the ``qmix.exp.RawData`` and ``qmix.exp.RawData0`` classes to control
+how experimental data is loaded and analyzed.
+
+Note:
+
+    This dictionary just contains the default values. You can overwrite these 
+    values by passing keyword arguments to ``RawData`` or ``RawData0``. For 
+    example, the default value for voltage units is ``"mV"``.
+    You can change this parameter to be ``"uV"`` by passing 
+    ``v_fmt="uV"`` to the ``RawData`` or ``RawData0`` class.
+
+All of the different parameters are described below along with their default 
+values.
 
 **Parameters:**
 
-    - CSV Files:
-        - The ``qmix.exp.RawData`` and ``qmix.exp.RawData0`` classes read in experimental data from CSV files. These parameters control how these files are loaded.
-        - ``delimiter = ","`` : The delimiter used in IV/IF CSV data files.
-        - ``usecols = (0,1)`` : Columns for voltage and current data in the IV/IF CSV data files.
-        - ``skip_header = 1`` : Number of rows to skip at the beginning of IV/IF CSV data files.
+    - CSV files:
+        - **Note:** All of the experimental data is expected to be stored in
+          CSV data files. These parameters control how the data is read in 
+          from these files.
+        - ``delimiter = ","`` : The delimiter used by the CSV data files.
+        - ``usecols = (0,1)`` : Which columns to import from the CSV data 
+          files.
+        - ``skip_header = 1`` : Number of rows to skip at the beginning CSV 
+          data files. Used to skip the header.
     - Units:
-        - ``v_fmt = "mv"`` : Units for voltage data (one of uV/mV/V).
-        - ``i_fmt = "mA"`` : Units for current data (one of uA/mA/A).
+        - ``v_fmt = "mV"`` : Units for imported voltage data. The options 
+          are: ``"uV"``, ``"mV"`` and ``"V"``.
+        - ``i_fmt = "mA"`` : Units for imported current data. The options
+          are: ``"uA"``, ``"mA"`` and ``"A"``.
     - Importing I-V data:
-        - ``vmax = 6e-3`` : Maximum voltage to import (in case the current is saturated beyond some bias voltage).
-        - ``npts = 6001`` : Number of points to use in IV data interpolation.
-    - Correct voltage/current offsets:
-        - ``ioffset = None`` : Offset in DC tunneling current data, in units A.
-        - ``voffset = None`` : Offset in DC bias voltage, in units V.
-        - ``voffset_range = 3e-4`` : Voltage range over which to look for the voltage offset, in units V.
-        - ``voffset_sigma = 1e-5`` : When finding voltage offset, assume this Gaussian width to find origin, in units V.
-    - Correct experimental I-V data:
-        - ``rseries = None`` : Correct for a series resistance using this resistance, in units ohms. Leave as ``None`` if there is no series resistance.
-        - ``iv_multiplier = 1.`` : Multiply the imported tunneling current by this number.
-    - Importing IF data:
-        - ``ifdata_vmax = 2.25`` : Maximum IF voltage to import (normalized to vgap).
-        - ``ifdata_npts = 3000`` : Number of points to use when interpolating IF data.
+        - ``vmax = 6e-3`` : Maximum voltage to import in units [V]. Used in
+          case the current is saturated beyond some bias voltage.
+        - ``npts = 6001`` : Number of points to use in the I-V data 
+          interpolation.
+        - ``debug = False`` : If set to ``True``, this will plot each step of
[…]
