### 2019-03-08 188620f
Add warning about how the Matplotlib axis is returned [ci skip]
