+++ qmix/exp/exp_data.py
@@ -189,7 +189,7 @@ class RawData0(object):
         message += "\tIleak: \t\t{:6.2f}\tuA\n".format(self.ileak * 1e6)
         message += "\n"
         message += "\tOffset:\t\t{:6.2f}\tmV\n".format(self.offset[0] * 1e3)
-        message += "\t       \t\t{:6.2f}\tuV\n".format(self.offset[1] * 1e6)
+        message += "\t       \t\t{:6.2f}\tuA\n".format(self.offset[1] * 1e6)
         message += "\n"
         message += "\tVint:  \t\t{:6.2f}\tmV\n".format(self.vint * 1e3)
 
@@ -214,8 +214,10 @@ class RawData0(object):
         resistance, subgap resistance, gap voltage, and gap current.
         
         Note: If ``fig_name`` is provided, this method will save the plot 
-        to the specified folder. If ``ax`` is provided, this method will return
-        the Matplotlib axis. **Do not define both.** 
+        to the specified folder and then close the plot. This means
+        that the Matplotlib axis object will not be returned in this
+        case. This is done to prevent too many plots from being open 
+        at the time.
         
         Args:
             fig_name (str): figure filename
@@ -289,8 +291,10 @@ class RawData0(object):
         """Plot DC I-V curve at the origin to see if there is an offset.
 
         Note: If ``fig_name`` is provided, this method will save the plot 
-        to the specified folder. If ``ax`` is provided, this method will return
-        the Matplotlib axis. **Do not define both.**
+        to the specified folder and then close the plot. This means
+        that the Matplotlib axis object will not be returned in this
+        case. This is done to prevent too many plots from being open 
+        at the time.
         
         Args:
             fig_name (str): figure filename
@@ -336,8 +340,10 @@ class RawData0(object):
         get resistance.
 
         Note: If ``fig_name`` is provided, this method will save the plot 
-        to the specified folder. If ``ax`` is provided, this method will return
-        the Matplotlib axis. **Do not define both.**
+        to the specified folder and then close the plot. This means
+        that the Matplotlib axis object will not be returned in this
+        case. This is done to prevent too many plots from being open 
+        at the time.
         
         Args:
             fig_name: figure filename
@@ -380,8 +386,10 @@ class RawData0(object):
         The static resistance is the DC voltage divided by the DC current.
 
         Note: If ``fig_name`` is provided, this method will save the plot 
-        to the specified folder. If ``ax`` is provided, this method will return
-        the Matplotlib axis. **Do not define both.**
+        to the specified folder and then close the plot. This means
+        that the Matplotlib axis object will not be returned in this
+        case. This is done to prevent too many plots from being open 
+        at the time.
         
         Args:
             fig_name: figure filename
@@ -429,8 +437,10 @@ class RawData0(object):
         The IF noise is calculated from the slope of the shot noise.
 
         Note: If ``fig_name`` is provided, this method will save the plot 
-        to the specified folder. If ``ax`` is provided, this method will return
-        the Matplotlib axis. **Do not define both.**
+        to the specified folder and then close the plot. This means
+        that the Matplotlib axis object will not be returned in this
+        case. This is done to prevent too many plots from being open 
+        at the time.
         
         Args:
             fig_name: figure filename
@@ -765,14 +775,17 @@ class RawData(object):
         else:
             cprint('\t- bad fit', 'WARNING')
         print("\t- embedding circuit:")
-        print("\t\t- voltage:      {:+6.2f}  x Vgap".format(vt_best))
-        print("\t\t- impedance:    {:+12.2f}  x Rn".format(zt_best))
+        print("\t\t- voltage:     \t{:+6.2f}\t\t* Vgap".format(vt_best))
+        print("\t\t- impedance:   \t{:+12.2f}\t* Rn".format(zt_best))
[…]
