### 2019-03-08 4dc3976
Add debugging features and the ability to multiply voltage data
