### 2019-03-08 cbe9e0e
Update package requirements for generating documentation [ci skip]
