### 2019-03-08 faa2b1e
Fix issue with conflicting packages
