+++ docs/source/index.rst
@@ -1,7 +1,7 @@
 QMix: Quantum Mixing Software
 =============================
 
-QMix is a software package for simulating the quasiparticle tunneling currents in Superconductor/Insulator/Superconductor (SIS) junctions. In radio astronomy, these junctions are used for heterodyne mixing at millimeter and submillimeter wavelengths. QMix can be used to simulate their performance, investigate their experimental results and optimize their operation.
+QMix is a software package for simulating the quasiparticle tunneling currents in Superconductor/Insulator/Superconductor (SIS) junctions. In radio astronomy, these junctions are used for heterodyne mixing at millimeter and submillimeter wavelengths. QMix can be used to simulate and optimize their performance, and investigate experimental results.
 
 In order to calculate the quasiparticle tunneling currents, QMix uses *multi-tone spectral domain analysis* (see :ref:`references` for more information). Among other applications, this makes QMix ideal for simulating power saturation, higher-order harmonics, sub-harmonic pumping and frequency multiplication.
 
@@ -11,8 +11,8 @@ Contents
 ========
 
 .. toctree::
-   :maxdepth: 1
-   :titlesonly:
+   :maxdepth: 3
+   :numbered:
    
    setup
    qmix
