### 2019-03-10 baf4c4b
Accept Numpy arrays into RawData and RawData0
