+++ qmix/exp/iv_data.py
@@ -1,10 +1,17 @@
-""" This sub-module contains functions for importing and analyzing 
-experimental I-V data measurements.
+"""This sub-module contains functions for importing and analyzing 
+experimental current-voltage (I-V) data.
 
 "I-V data" is the DC tunneling current versus DC bias voltage that is measured
-from the SIS device. In general, I use the term "DC I-V data" for I-V data
-that is collected with no LO present, and "I-V data" for I-V data that is
-collected with the LO present (also known as the "pumped I-V curve").
+from an SIS junction. In general, the term "DC I-V data" is used for I-V data
+that is collected with no local-oscillator (LO) injection, and "I-V data" is
+used for I-V data that is collected with LO injection (also known as the 
+"pumped I-V curve").
+
+Note:
+    
+    The I-V data is expected either in the form of a CSV file or a Numpy 
+    array. Either way the data should have two columns: the first for voltage 
+    and the second for current.
 
 """
 
@@ -17,19 +24,17 @@ import scipy.constants as sc
 from scipy.signal import savgol_filter
 
 from .clean_data import remove_doubles_xy, remove_nans_xy, sort_xy
+from qmix.exp.parameters import params as PARAMS
 from qmix.mathfn.filters import gauss_conv
 from qmix.mathfn.misc import slope
 
 filterwarnings(action="ignore", module="scipy", message="^internal gelsd")
 
-# Voltage units
-_vfmt_dict = {'uV': 1e-6, 'mV': 1e-3, 'V': 1}
-
-# Current units
-_ifmt_dict = {'uA': 1e-6, 'mA': 1e-3, 'A': 1}
+_vfmt_dict = {'uV': 1e-6, 'mV': 1e-3, 'V': 1}  # Voltage units
+_ifmt_dict = {'uA': 1e-6, 'mA': 1e-3, 'A': 1}  # Current units
 
 
-# Import and analyze i-v data ------------------------------------------------
+# Import and analyze I-V data ------------------------------------------------
 
 DCIVData = namedtuple('DCIVData', ['vraw', 'iraw', 'vnorm', 'inorm', 'vgap',
                                    'igap', 'fgap', 'rn', 'rsg', 'offset',
@@ -38,53 +43,76 @@ DCIVData.__doc__ = """\
 Struct for DC I-V curve metadata.
 
 Args:
-    vraw (ndarray): DC bias voltage in units V. This data has filtered and the
-        offset has been corrected.
-    iraw (ndarray): DC tunneling current in units A. This data has filtered 
+    vraw (ndarray): DC bias voltage in units [V]. This data has been filtered 
         and the offset has been corrected.
-    vnorm (ndarray): DC bias voltage (normalized).
-    inorm (ndarray): DC tunneling current (normalized).
-    vgap (float): Gap voltage in units V.
-    igap (flaot): Gap current in units A.
-    fgap (float): Gap frequency in units Hz.
-    rn (float): Normal-state resistance in units ohms.
-    rsg (float): Sub-gap resistance in units ohms.
-    offset (tuple): Voltage and current offset in the raw measured data.
+    iraw (ndarray): DC tunneling current in units [A]. This data has been 
+        filtered and the offset has been corrected.
+    vnorm (ndarray): DC bias voltage, normalized to the gap voltage.
+    inorm (ndarray): DC tunneling current, normalized to the gap current.
+    vgap (float): Gap voltage in units [V].
+    igap (flaot): Gap current in units [A].
+    fgap (float): Gap frequency in units [Hz].
+    rn (float): Normal-state resistance in units [ohms].
+    rsg (float): Sub-gap resistance in units [ohms].
+    offset (tuple): Voltage and current offset in the raw measured data, in
+        units [V] and [A], respectively.
[…]
