+++ qmix/exp/if_data.py
@@ -2,7 +2,15 @@
 experimental IF power measurements.
 
 The "IF data" is the IF output power from the SIS device versus bias voltage.
+The term "DC IF data" is used for IF power with no LO injection, and "IF data"
+is used for IF power with LO injection.
 
+Note:
+    
+    The IF data is expected either in the form of a CSV file or a Numpy 
+    array. Either way the data should have two columns: the first for voltage 
+    and the second for current.
+    
 """
 
 from collections import namedtuple
@@ -13,12 +21,11 @@ from scipy.signal import savgol_filter
 import scipy.constants as sc
 
 from qmix.exp.clean_data import remove_doubles_matrix, remove_nans_matrix
+from qmix.exp.parameters import params as PARAMS
 from qmix.mathfn import slope_span_n
 from qmix.mathfn.filters import gauss_conv
 
-
-# Voltage units
-_vfmt_dict = {'uV': 1e-6, 'mV': 1e-3, 'V': 1}
+_vfmt_dict = {'uV': 1e-6, 'mV': 1e-3, 'V': 1}  # Voltage units
 
 
 # Load IF data and determine noise temperature -------------------------------
@@ -38,58 +45,93 @@ Args:
 """
 
 
-def dcif_data(dcif_file, dc, **kwargs):
+def dcif_data(ifdata, dc, **kwargs):
     """Analyze DC IF measurements.
 
     This is the IF data that is measured with no LO present. This data is
     used to analyze the shot noise, which can then be used to convert the IF 
     data into units 'K' and estimate the IF noise component.
-
+            
     Args:
-        dcif_file (str): DC IF filename (no LO present)
-        dc (qmix.exp.iv_data.DCIVData): DC I-V data structure
-        **kwargs: keyword arguments
+        ifdata: IF data. Either a CSV data file or a Numpy array. The data
+            should have two columns: the first for voltage, and the second
+            for IF power. To pass a Numpy array, set the ``input_type`` 
+            keyword argument to ``"numpy"``. To pass a CSV data file, set the 
+            ``input_type`` keyword argument to ``"csv"``. The properties of 
+            the CSV file can be set through additional keyword arguments.
+            (See below).
+        dc (qmix.exp.iv_data.DCIVData): DC I-V metadata.
 
-    Keyword arguments:
-        v_fmt (str): voltage units ('mV', 'V', etc.)
-        vmax (str): maximum voltage (in case the data is saturated above some
-            value)
+    Keyword Args:
+        input_type (str): Input type ('csv' or 'numpy').
+        delimiter (str): Delimiter for CSV files.
+        usecols (tuple): List of columns to import (tuple of length 2).
+        skip_header (int): Number of rows to skip, used to skip the header.
+        v_fmt (str): Units for voltage ('uV', 'mV', or 'V').
+        i_fmt (str): Units for current ('uA', 'mA', or 'A').
+        rseries (float): Series resistance in experimental measurement 
+            system, in units [ohms].
+        v_multiplier (float): Multiply the imported voltage by this value.
+        ifdata_vmax (float): Maximum IF voltage to import.
+        ifdata_npts (int): Number of points for interpolation.
+        ifdata_sigma (float): Std. dev. of Gaussian used for filtering.
+        vshot (list): Range of voltages for fitting shot noise slope.
+        verbose (bool): Print to terminal.
 
     Returns:
[…]

[…]
