### 2019-03-10 daa5a66
Minor changes to docstrings [ci skip]
