+++ qmix/circuit.py
@@ -6,8 +6,8 @@ circuit.
     In experimental systems, SIS junctions are embedded within complex RF 
     networks. These networks are referred to as the embedding circuit. Since
     all of the components in these embedding circuits are **linear**, the 
-    embedding circuit can be reduced to a **Thevenin equivalent circuit** for
-    **each tone and harmonic.**
+    embedding circuit can be reduced to a **Thevenin equivalent circuit**, with
+    one for **each tone and harmonic.**
     
     To fully describe the embedding circuit, 3 bits of information are needed
     for each signal that is applied to the junction:
@@ -53,7 +53,7 @@ class EmbeddingCircuit(object):
         set manually. In this way, this class is sort of like a fancy struct.
         The class attributes that have to be set manually are:
 
-           - ``vph``:  photon voltage (freq / gap freq)
+           - ``vph``:  photon voltage (normalized to the gap voltage)
            - ``vt``:  Thevenin voltage (normalized to the gap voltage)
            - ``zt``:  Thevenin impedance (normalized to the normal resistance)
 
@@ -62,38 +62,33 @@ class EmbeddingCircuit(object):
 
     Example:
 
-        To create an instance of the embedding circuit class with 2 tones and
-        3 harmonices:
+        Here we will create an instance of the embedding circuit class with
+        2 tones and 3 harmonics:
 
         >>> cct1 = EmbeddingCircuit(2, 3, name='Test 1')
         >>> print(cct1)
         Embedding circuit (Tones:2, Harmonics:3): Test 1
 
-        Once initialized, you can begin defining the properties of the
+        Once initialized, we can begin defining the properties of the
         embedding circuit. I normally start with the photon voltages (or,
         equivalently, the normalized frequencies):
 
         >>> cct1.vph[1] = 0.30  # first tone
         >>> cct1.vph[2] = 0.32  # second tone
 
-        Then, you have to set the voltages and impedances for all of the
+        Then, we have to set the voltages and impedances for all of the
         different signals. For example, for the 1st harmonic of the 1st tone:
 
         >>> cct1.vt[1,1] = 0.5             # Thevenin voltage
         >>> cct1.zt[1,1] = 0.3 + 1j * 0.1  # Thevenin impedance
 
-        If you would like the ability to use non-normalized values (e.g., set
-        the available power of a signal in units [W]), you need to define the
-        properties of the junction during initialization. For example:
+        This has to be done for each signal (a total of 6 times in this case).
+
+        In order to use non-normalized values (e.g., set the available power of
+        a signal in units [W]), we need to define the electrical properties of
+        the junction during initialization. For example:
 
         >>> cct2 = EmbeddingCircuit(1, 1, vgap=2.8e-3, rn=14.)
-
-        Note that you only need to define either the gap voltage or the gap
-        frequency since they can be calculated from each other.
-
-        >>> fgap = cct2.fgap
-        >>> round(fgap / 1e9)  # to display in [GHz]
-        677
 
         You can now set the photon voltage using the frequency of the applied
         signal. E.g.:
@@ -115,17 +110,19 @@ class EmbeddingCircuit(object):
         -50.0
 
     Args:
-        num_f (int, optional, default is 1): Number of fundamental frequencies (tones)
-            applied to the junction.
-        num_p (int, optional, default is 1): Number of harmonics included for each tone.
+        num_f (int, optional, default is 1): Number of fundamental frequencies
+            (tones) applied to the junction.
+        num_p (int, optional, default is 1): Number of harmonics included for
+            each tone.
[…]
