### 2019-03-10 0fa39dc
Automatically identify input data type
