+++ qmix/exp/iv_data.py
@@ -71,14 +71,11 @@ def dciv_curve(ivdata, **kwargs):
     Args:
         ivdata: DC I-V data. Either a CSV data file or a Numpy array. The data
             should have two columns: the first for voltage, and the second
-            for current. To pass a Numpy array, set the ``input_type`` keyword
-            argument to ``"numpy"``. To pass a CSV data file, set the 
-            ``input_type`` keyword argument to ``"csv"``. The properties of 
+            for current. If you are using CSV files, the properties of 
             the CSV file can be set through additional keyword arguments.
             (See below).
 
     Keyword Args:
-        input_type (str): Input type ('csv' or 'numpy').
         delimiter (str): Delimiter for CSV files.
         usecols (tuple): List of columns to import (tuple of length 2).
         skip_header (int): Number of rows to skip, used to skip the header.
@@ -206,16 +203,13 @@ def iv_curve(ivdata, dc, **kwargs):
     Args:
         ivdata: I-V data. Either a CSV data file or a Numpy array. The data
             should have two columns: the first for voltage, and the second
-            for current. To pass a Numpy array, set the ``input_type`` keyword
-            argument to ``"numpy"``. To pass a CSV data file, set the 
-            ``input_type`` keyword argument to ``"csv"``. The properties of 
-            the CSV file can be set through additional keyword arguments.
-            (See below).
+            for current. If you are using a CSV file, the properties of 
+            the CSV file can be set through additional keyword arguments
+            (see below).
         dc (qmix.exp.iv_data.DCIVData): DC I-V data metadata. Generated 
             previously by ``dciv_curve``.
 
     Keyword Args:
-        input_type (str): Input type ('csv' or 'numpy').
         delimiter (str): Delimiter for CSV files.
         usecols (tuple): List of columns to import (tuple of length 2).
         skip_header (int): Number of rows to skip, used to skip the header.
@@ -327,7 +321,6 @@ def _load_iv(ivdata, **kw):
             for current.
 
     Keyword Arguments:
-        input_type: input type ('csv' or 'numpy')
         v_fmt: voltage units ('uV', 'mV', 'V')
         i_fmt: current units ('uA', 'mA', 'A')
         usecols: list of columns to use (tuple of length 2)
@@ -341,24 +334,23 @@ def _load_iv(ivdata, **kw):
 
     # Unpack keyword arguments
     skip_header = kw.get('skip_header', PARAMS['skip_header'])
-    input_type = kw.get('input_type', PARAMS['input_type'])
     delimiter = kw.get('delimiter', PARAMS['delimiter'])
     usecols = kw.get('usecols', PARAMS['usecols'])
     v_fmt = kw.get('v_fmt', PARAMS['v_fmt'])
     i_fmt = kw.get('i_fmt', PARAMS['i_fmt'])
 
     # Import raw I-V data
-    if input_type.lower() == 'csv':
+    if isinstance(ivdata, str):  # input: CSV file
         vraw, iraw = np.genfromtxt(ivdata, delimiter=delimiter,
                                    usecols=usecols, skip_header=skip_header).T
-    elif input_type.lower() == 'numpy':
+    elif isinstance(ivdata, np.ndarray):  # input: Numpy array
         assert isinstance(ivdata, np.ndarray), \
             'I-V data should be a Numpy array.'
         assert ivdata.ndim == 2, 'I-V data should be 2-dimensional.'
         assert ivdata.shape[1] == 2, 'I-V data should have 2 columns.'
         vraw, iraw = ivdata.T
     else:
-        raise ValueError("Input type not recognized.")
+        raise ValueError("Input data type not recognized.")
 
     # Set units
     volt_v = vraw * _vfmt_dict[v_fmt]

[…]
