+++ _sources/analyze-experimental-data.ipynb.txt
@@ -6,14 +6,16 @@
    "source": [
     "# Example #3: Analyzing experimental data\n",
     "\n",
-    "This notebook will show a brief example of how QMix can be used to analyze experimental data. Note that a wide range of parameters can be passed to these functions in order to import data that has been stored in different formats. Please see [qmix.exp.parameters](https://garrettj403.github.io/QMix/qmix.exp.html#module-qmix.exp.parameters) for a full list of optional keyword arguments.\n",
+    "This notebook will show a brief example of how QMix can be used to analyze experimental data. \n",
     "\n",
-    "For the purposes of this notebook, example data has been included in a directory called ``eg-230-data/``. This is experimental data that was taken from an SIS device illuminated by a local-oscillator at 230 GHz (Garrett 2018).\n",
+    "For the purposes of this notebook, example data has been included in a directory called ``eg-230-data/``. This is experimental data that was taken from an SIS device illuminated by a local-oscillator at 230 GHz (see Garrett 2018).\n",
     "\n",
-    "QMix includes two classes that can be used to analyze experimental data:\n",
+    "QMix includes two classes that can be used to analyze this experimental data:\n",
     "\n",
-    "   - [qmix.exp.RawData0](https://garrettj403.github.io/QMix/qmix.exp.html#qmix.exp.exp_data.RawData0), which is used for DC data (no LO illumination), and\n",
-    "   - [qmix.exp.RawData](https://garrettj403.github.io/QMix/qmix.exp.html#qmix.exp.exp_data.RawData), which is used to analyze pumped data (with LO illumination)."
+    "   - [qmix.exp.RawData0](https://garrettj403.github.io/QMix/qmix.exp.html#qmix.exp.exp_data.RawData0), which is used to analyze DC data (no LO illumination), and\n",
+    "   - [qmix.exp.RawData](https://garrettj403.github.io/QMix/qmix.exp.html#qmix.exp.exp_data.RawData), which is used to analyze pumped data (with LO illumination).\n",
+    "   \n",
+    "Note that a wide range of parameters can be passed to these classes in order to control how the data is imported and analyzed. Please see [qmix.exp.parameters](https://garrettj403.github.io/QMix/qmix.exp.html#module-qmix.exp.parameters) for a full list of optional keyword arguments and their default values."
    ]
   },
   {
@@ -33,7 +35,7 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "## DC data\n",
+    "## Analyzing DC data\n",
     "\n",
     "This is the data collected from the SIS device with **no local-oscillator (LO) illumination.**\n",
     "\n",
@@ -105,7 +107,7 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "In order to import these files properly, we will define the properties of the CSV files (which will be passed on to the ``RawData0`` and ``RawData`` classes)."
+    "In order to import these files properly, we will define the properties of the CSV files, which will then be passed on to the ``RawData0`` class."
    ]
   },
   {
@@ -114,11 +116,11 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "csv_params = dict(delimiter = ',',  # delimiter used by CSV files\n",
-    "                  usecols = (0,1),  # columns to import\n",
-    "                  skip_header = 1,  # skip the first row (the header)\n",
-    "                  v_fmt = 'mV',     # units for voltage data\n",
-    "                  i_fmt = 'mA')     # units for current data"
+    "csv_params = dict(delimiter = ',',     # delimiter used by CSV files\n",
+    "                  usecols = (0,1),     # columns to import\n",
[…]
