### 2019-03-11 #1231 @garrettj403
Hi @FaustinCarter, thank you for your detailed review.

I think that I have addressed all of your points now.

> ## General documentation
>
> There are a lot of helper functions (resp_swap, for instance) that do specific things, but I don't immediately see why they would be needed. I expect that this is because I am not an SIS mixer expert, but I imagine some new grad student might also not appreciate them fully. It would be good to include at least a minimal use case for all helper functions (i.e. this is useful for doing the following thing). If it is a function the user is not ever really supposed to use (i.e. just for internal Class calculations), consider prepending the function name with an underscore

- I added more use cases for the obscure methods in [``qmix.respfn.RespFn``](https://garrettj403.github.io/QMix/qmix.html#qmix.respfn.RespFn) (e.g., [``resp_conj``](https://garrettj403.github.io/QMix/qmix.html#qmix.respfn.RespFn.resp_conj)).
- The methods in [``qmix.circuit.EmbeddingCircuit``](https://garrettj403.github.io/QMix/qmix.html#qmix.circuit.EmbeddingCircuit) are a bit more straight-forward, but I added use cases to a couple of the methods (e.g., [``lock``](https://garrettj403.github.io/QMix/qmix.html#qmix.circuit.EmbeddingCircuit.lock)).

> There are several different response function types. It might be useful to say something about when/why each one ought to be used.

- I added more information to the [``qmix.respfn.RespFn``](https://garrettj403.github.io/QMix/qmix.html#module-qmix.respfn) module's docstring to describe the different types of response functions.

> There is a lot of SIS specific jargon that I'm not familiar with. One example is the term "photon voltage". Is that the photon energy divided by electron charge, like it sounds, or something else? Probably this is all well understood by field experts, so I'll defer to the field expert as to whether everything was understandable on the first go. You might consider asking a newish grad student to read through your docs and point out all the terms they are unfamiliar with, but I'm not going to hold you to it!

- The [``qmix.circuit.EmbeddingCircuit``](https://garrettj403.github.io/QMix/qmix.html#qmix.circuit.EmbeddingCircuit) class contains the most SIS jargon, so I now describe all of the class attributes in detail (e.g., photon voltage: [``vph``](https://garrettj403.github.io/QMix/qmix.html#qmix.circuit.EmbeddingCircuit.vph))). Also, the documentation for [``qmix.circuit.EmbeddingCircuit``](https://garrettj403.github.io/QMix/qmix.html#qmix.circuit.EmbeddingCircuit) now includes an example of how to initialize and setup the class.

> One of the submodules in your exp sub-package is just a dictionary of default parameter values. Although the comments in the code give some explanation, the auto-doc is very garbled. This needs to be fixed. Also, in Examples 1 and 2, there should be some explicit reference to this default parameter list, or all the values for the toy models should be explicitly passed to the constructors. The less that happens under the hood the better for Examples.

- I fixed the documentation for [``qmix.exp.parameters.params``](https://garrettj403.github.io/QMix/qmix.exp.html#module-qmix.exp.parameters). There isn't a good way to document dictionaries, so I just describe the different parameters in the module's docstring.
- I refer to the parameters file [here in Example #3](https://garrettj403.github.io/QMix/analyze-experimental-data.html#Example-#3:-Analyzing-experimental-data), and I also do more to explain the different keyword arguments and how they are passed to ``RawData0`` [here](https://garrettj403.github.io/QMix/analyze-experimental-data.html#Properties-of-the-experimental-DC-data), on lines 4-8.

> ## Handling user data
>
> In the experimental data sub-package, there is not a clear specification for what data format is required for both RawData0 and RawData classes. There is a link in the docs to some example data files, but those files do not even have column names, so making inferences is hard.
>
> I recommend the following fixes:
>
>     * add column names to the csv files, or at least include them in the readme file (along with expected units)
>
>     * Explicitly spell out what the expected format is for any data files that the user might want to read in.
>

- I've added column names to all of the CSV files, and I have included a detailed description of the file format [here in Example #3](https://garrettj403.github.io/QMix/analyze-experimental-data.html#Properties-of-the-experimental-DC-data).

> Optionally (and for future reference), I think it is good practice to accept a data structure rather than a file into your routines. For instance, if you accept a numpy.ndarray or a pandas.DataFrame or xarray.Dataset or a python dict of iterables, you would give the user the freedom to load up their data in whatever way they like. You can optionally provide a parser that converts your preferred type of files into that data structure. This saves folks from having to load up their data, save their data in your format, and then reload it from file when they run your code.

- I've changed ``RawData`` and ``RawData0`` so that they now accept data in the form of Numpy arrays (in addition to CSV files). You can see an example [here, in Example #3](https://garrettj403.github.io/QMix/analyze-experimental-data.html#Properties-of-the-experimental-data), in lines 12-16.

> ## Plotting
>
> All of your plotting routines should return a handle to the matplotlib figure object at a minimum. It is very likely that a user will want to continue to annotate, or plot extra stuff, or take your plot and stick it into a larger figure, and they can't do any of this easily with a saved pdf.
>
> The typical way of handling this is to accept a handle to a matplotlib axis as an optional keyword argument and then:
>
>     * if they don't pass in their own axis, just create a new figure, plot stuff up, and return the figure.
>
>     * if they do pass in their own axis, plot directly on the axis passed in
>
>
> There are also probably other sane variations on this theme, and I'll happily accept any of them, as long as the user has some way of getting a matplotlib object back from the plotting routines.

- All plotting functions now accept and return Matplotlib axis objects. You can see an example [here, on line 19](https://garrettj403.github.io/QMix/analyze-experimental-data.html#Impedance-recovery).

Thank you again and please let me know if I missed anything!
