+++ _sources/analyze-experimental-data.ipynb.txt
@@ -528,9 +528,9 @@
    "outputs": [
     {
      "data": {
-      "image/png": "iVBORw0KGgoAAAANSUhEUgAAAcYAAAFECAYAAAC9Gp76AAAABHNCSVQICAgIfAhkiAAAAAlwSFlzAAALEgAACxIB0t1+/AAAADl0RVh0U29mdHdhcmUAbWF0cGxvdGxpYiB2ZXJzaW9uIDMuMC4yLCBodHRwOi8vbWF0cGxvdGxpYi5vcmcvOIA7rQAAIABJREFUeJzsXXe8HFXZft7dvSWdFJIASYAQUKRDRBSUYiiCAkoTLCAfAlIsVGmCNCkqIoIYFRUFFEUE6aACgoKEJp0AhhBaEhLSc+/dnff7Y/fefc/svLNndmZnyz0Pv/yYnT1t987OM8/bDjEzHBwcHBwcHIrINHoBDg4ODg4OzQRHjA4ODg4ODgKOGB0cHBwcHAQcMTo4ODg4OAg4YnRwcHBwcBBwxOjg4ODg4CDgiNHBwcHBwUHAEaODg4ODg4OAI0YHBwcHBweBXKMXUA+MG5Pl9SZ3hLaxq/dTvVWSdYM4wdHqVc+ofuNGHNmyefT1UuQeadeOamStqrhzcw3fb72R7PdZn88n1/jyMz0LmXnNpMbefedh/N6iQqQ+j/+3525m3iOpNTQb2pIY15vcgf/cPSW0jQev6jgFrt7Gs/xZ2c1nO1b1doWIP3fbz2GzxuqfNGDciOstWDaPupZCDTc2L+KdtZY5kurvcePmLvav3UgVd+0a4n4mYyyujxHOE9/bTuvPfj3Jsd9bVKh6v/Qju9bscUmuodnQlsToYAdbMqwX6kWG7YhWJcM4RAgkS4ZJEqAxbgpkWE8w7B7cBxMcMTo4ODgMarCVdWwwwRGjQ1XUy3xqN3e09rWZcaMpiaim01rmiKtunEosIimV2OqqMAxFxTiIzTEBcMRYA9L2K1r7/yzaNfIHENV0agtHhkU4MizCkWF0OFOqCUeMPjiTQhGtpBKtx21CMoyLOITSyEAaIDkybLXgmWYDg60D/wYLHDE6ODg4DHI4U6qJQUWM7WguqJd5Mik0k/m0XohngqwlPWRwqsTBrAzraYlgNP99JG0MKmJMA2n7FW1glfdonUNZH9iYT5vVlxgVrUCGcQmwH43OmzTGcmSowilGE44YLeEuHHvYPn3Ww5fYrAn60Qk6fUJpluT7ZiTDdiA/DQz7B+PBAkeMJRQKjP882YMFi7LggB+5LTHalDbzElSMNq1s2tiqMZt2XunH3tnRh40/RBg/Pms5uoODw2AAEZ0HYB8UbynzARzGzG8FtDsUwJmll+cz82/SWJ8jRgDMjJvvKGBVYVu8Nmcl/nb/y5g4YQTWXHM4/u/Qj/a3quj3+JNv4LXXFuKA/bYqj6XM8dbbS/CnPz+Frx+740Cb08/6Ky487zO44qoHcPwxO1rXC/UAXHnVgzj2mE+Efaqq40h+PvM7t+H8cz9d40jBRL58eQ9u+csD2HufhZgwoTZyrJfp1qnE5okqBWL6aZ2JNDYa4LO
[…]
