### 2019-03-11 693c30a
Minor changes to documentation [ci skip]
