+++ paper/paper.bib
@@ -8,12 +8,12 @@ year = 2018,
 
 @article{Garrett2019,
   author       = {J. D. Garrett and G. Yassin},
-  title        = {{garrettj403/QMix  (Version v1.0.1)}},
-  month        = feb,
+  title        = {{garrettj403/QMix  (Version v1.0.2)}},
+  month        = mar,
   year         = 2019,
   publisher    = {Zenodo},
-  doi          = {10.5281/zenodo.2538163},
-  url          = {http://doi.org/10.5281/zenodo.2557839}
+  doi          = {10.5281/zenodo.2590165},
+  url          = {http://doi.org/10.5281/zenodo.2590165}
 }
 
 @phdthesis{Kittara2002,
