+++ CHANGES.md
@@ -1,5 +1,5 @@
-v1.0.2 (unreleased)
-===================
+v1.0.2 (11-Mar-2019)
+====================
 
 New Features
 ------------
