+++ setup.cfg
@@ -7,5 +7,5 @@ author_email = garrettj403@gmail.com
 license = GPL v3
 url = https://github.com/garrettj403/QMix
 github_project = garrettj403/QMix
-install_requires = cycler, matplotlib, numba, numpy, scipy
-version = 1.0.0
+install_requires = matplotlib, numba, numpy, scipy
+version = 1.0.2
