### 2019-03-11 #1231 @PaulKGrimes
Hi Lorena,

Yes, I'm ready to recommend publication.

Paul

On Sun, Mar 10, 2019, 15:56 Lorena A. Barba <notifications@github.com>
wrote:

> @PaulKGrimes <https://github.com/PaulKGrimes> — I see you've checked all
> your review items. Are you ready to recommend publication?
>
