### 2019-03-20 #1231 @labarba
After you do this ☝️, the next step is to create a tagged release (post the version number here), and update your Zenodo deposit to a new version (no need for a new DOI, just a version). Post the archive DOI here.
