### 2019-03-20 aa55318
Remove last sentence (GitHub/Zenodo reference) [ci skip]
