+++ paper/paper.bib
@@ -6,16 +6,6 @@ type = {DPhil},
 year = 2018,
 }
 
-@article{Garrett2019,
-  author       = {J. D. Garrett and G. Yassin},
-  title        = {{garrettj403/QMix  (Version v1.0.2)}},
-  month        = mar,
-  year         = 2019,
-  publisher    = {Zenodo},
-  doi          = {10.5281/zenodo.2590165},
-  url          = {http://doi.org/10.5281/zenodo.2590165}
-}
-
 @phdthesis{Kittara2002,
 author = {P. Kittara},
 school = {University of Cambridge},
