+++ setup.cfg
@@ -8,4 +8,4 @@ license = GPL v3
 url = https://github.com/garrettj403/QMix
 github_project = garrettj403/QMix
 install_requires = matplotlib, numba, numpy, scipy
-version = 1.0.2
+version = 1.0.3
