### 2019-03-20 #1231 @garrettj403
Hi @labarba, I've removed the last sentence from ``paper.md`` and issued a new release. The new version number is [v1.0.3](https://github.com/garrettj403/QMix/releases/tag/v1.0.3) and the Zenodo DOI is [10.5281/zenodo.2538162](https://doi.org/10.5281/zenodo.2538162) (for the archive as a whole).
