### 2019-03-20 #1231 @labarba
Please edit the metadata of your Zenodo deposit so the title and author list match the JOSS paper.
