### 2019-03-20 #1231 @garrettj403
Okay, I've updated the [Zenodo archive](https://zenodo.org/record/2600288). The title and authors now match the JOSS paper.
