### 2019-03-20 #1231 @labarba
Congratulations, @garrettj403, your JOSS paper is published 👏

Huge thanks to the reviewers: @FaustinCarter, @josephhardinee, @PaulKGrimes 🙏
